### 2023-09-26 #5878 @editorialbot
[PRE REVIEW]: DeepRank2: Mining 3D Protein Structures with Geometric Deep Learning
**Submitting author:** @gcroci2 (<a href="http://orcid.org/0000-0002-0823-0121">Giulia Crocioni</a>)
**Repository:** https://github.com/DeepRank/deeprank2
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
