+++ tests/test_query.py
@@ -6,6 +6,7 @@ from typing import List
 import h5py
 import numpy as np
 import pytest
+from pdb2sql import pdb2sql
 
 from deeprank2.dataset import GraphDataset, GridDataset
 from deeprank2.domain import edgestorage as Efeat
@@ -16,9 +17,10 @@ from deeprank2.domain.aminoacidlist import (alanine, arginine, asparagine,
                                             glycine, leucine, lysine,
                                             phenylalanine)
 from deeprank2.features import components, conservation, contact, surfacearea
-from deeprank2.molstruct.residue import Residue
+from deeprank2.molstruct.residue import Chain, Residue
 from deeprank2.query import (ProteinProteinInterfaceQuery, QueryCollection,
                              SingleResidueVariantQuery)
+from deeprank2.utils.buildgraph import get_structure
 from deeprank2.utils.graph import Graph
 from deeprank2.utils.grid import GridSettings, MapMethod
 
@@ -174,8 +176,9 @@ def test_variant_graph_101M():
     )
 
 def test_variant_in_atomic_graph():
+    pdb_path = "tests/data/pdb/101M/101M.pdb"
     query = SingleResidueVariantQuery(
-        pdb_path="tests/data/pdb/101M/101M.pdb",
+        pdb_path=pdb_path,
         resolution="atom",
         chain_ids="A",
         variant_residue_number=27,
@@ -190,28 +193,26 @@ def test_variant_in_atomic_graph():
 
     g = query.build([components])
 
+    pdb = pdb2sql(pdb_path)
+    try:
+        structure = get_structure(pdb, "101M")
+    finally:
+        pdb._close() # pylint: disable=protected-access
+
+    chain = Chain(structure, 'A')
+
     variant_residue = Residue(
-        chain = '101M ' + query.chain_ids[0],
+        chain = chain,
         number = query.variant_residue_number,
         amino_acid = query.wildtype_amino_acid,
         insertion_code = query.insertion_code,
     )
     all_residues = list(set(node.id.residue for node in g.nodes))
-    print (variant_residue)
-    # print (type(variant_residue), type(all_residues[0]))
-    # for res in all_residues:
-    #     print (res)
-    print (all_residues[4])
-    # assert variant_residue in all_residues
+    print (1, variant_residue)
+    print (2, all_residues[4])
+    assert variant_residue in all_residues
+    assert False
 
-    another_residue = Residue(
-        chain='101M A',
-        number=27,
-        amino_acid=phenylalanine,
-        insertion_code=None,
-    )
-
-    assert variant_residue == another_residue
 
 def test_variant_graph_1A0Z():
     query = SingleResidueVariantQuery(
