### 2023-10-09 6480644
make distance cutoff uniform for query types
