### 2023-10-09 c2151e7
rename `DeepRankQuery` back to `Query`
