+++ tests/test_querycollection.py
@@ -12,7 +12,7 @@ from deeprank2.domain import edgestorage as Efeat
 from deeprank2.domain import nodestorage as Nfeat
 from deeprank2.domain.aminoacidlist import alanine, phenylalanine
 from deeprank2.features import components, contact, surfacearea
-from deeprank2.query import (DeepRankQuery, ProteinProteinInterfaceQuery,
+from deeprank2.query import (ProteinProteinInterfaceQuery, Query,
                              QueryCollection, SingleResidueVariantQuery)
 from deeprank2.tools.target import compute_ppi_scores
 
@@ -131,7 +131,7 @@ def test_querycollection_process():
         assert isinstance(collection.queries, list)
         assert len(collection.queries) == n_queries
         for query in collection.queries:
-            assert issubclass(type(query), DeepRankQuery)
+            assert issubclass(type(query), Query)
 
         rmtree(output_directory)
 
@@ -224,9 +224,8 @@ def test_querycollection_process_combine_output_false():
 
 
 def test_querycollection_duplicates_add():
-    """
-    Tests add method of QueryCollection class.
-    """
+    """Tests add method of QueryCollection class."""
+
     ref_path = "tests/data/ref/1ATN/1ATN.pdb"
     pssm_path1 = "tests/data/pssm/1ATN/1ATN.A.pdb.pssm"
     pssm_path2 = "tests/data/pssm/1ATN/1ATN.B.pdb.pssm"
