### 2023-10-10 de72dd6
Update README.md
Add pip install commands for the dependencies listed in the README
