### 2023-10-19 a5ad04a
add relevant attributes to the Trainer and improve their logic
