+++ deeprank2/trainer.py
@@ -3,6 +3,7 @@ import logging
 from time import time
 from typing import List, Optional, Tuple, Union
 
+import dill
 import h5py
 import numpy as np
 import torch
@@ -119,7 +120,10 @@ class Trainer():
             _log.info(f"CUDA device name is {torch.cuda.get_device_name(0)}.")
             _log.info(f"Number of GPUs set to {self.ngpu}.")
 
-        if pretrained_model is None:
+        self.pretrained_model_path = pretrained_model
+        self.model_load_state_dict = None
+
+        if self.pretrained_model_path is None:
             if self.dataset_train is None:
                 raise ValueError("No training data specified. Training data is required if there is no pretrained model.")
             if self.neuralnet is None:
@@ -166,10 +170,7 @@ class Trainer():
                 raise ValueError("No neural network class found. Please add it to complete loading the pretrained model.")
             if self.dataset_test is None:
                 raise ValueError("No dataset_test found. Please add it to evaluate the pretrained model.")
-            if self.target is None:
-                raise ValueError("No target set. Make sure the pretrained model explicitly defines the target to train against.")
 
-            self.pretrained_model_path = pretrained_model
             self.classes_to_index = self.dataset_test.classes_to_index
 
             self._load_params()
@@ -223,17 +224,25 @@ class Trainer():
             self.node_features = dataset.node_features
             self.edge_features = dataset.edge_features
             self.features = None
+            self.features_transform = dataset.features_transform
+            self.means = dataset.means
+            self.devs = dataset.devs
 
         elif isinstance(dataset, GridDataset):
             self.clustering_method = None
             self.node_features = None
             self.edge_features = None
             self.features = dataset.features
+            self.features_transform = None
+            self.means = None
+            self.devs = None
         else:
             raise TypeError(type(dataset))
 
         self.target = dataset.target
+        self.target_transform = dataset.target_transform
         self.task = dataset.task
+        self.classes = dataset.classes
 
     def _load_model(self):
         """Loads the neural network model."""
@@ -549,6 +558,9 @@ class Trainer():
         else:
             self.valid_loader = None
             _log.info("No validation set provided\n")
+            _log.warning(
+                "Training data will be used both for learning and model selection, which may lead to overfitting." +
+                "\nIt is usually preferable to use a validation set during the training phase.")
 
         # Assign weights to each class
         if self.task == targets.CLASSIF and self.class_weights:
@@ -622,9 +634,6 @@ class Trainer():
                     # if no validation set, save the best performing model on the training set
                     if best_model:
                         if min(train_losses) == loss_:
-                            _log.warning(
-                                "Training data is used both for learning and model selection, which will to overfitting." +
-                                "\n\tIt is preferable to use an independent training and validation data sets.")
                             checkpoint_model = self._save_model()
                             self.epoch_saved_model = epoch
                             _log.info(f'Best model saved at epoch # {self.epoch_saved_model}.')
@@ -637,7 +646,7 @@ class Trainer():
 
         # Now that the training loop is over, save the model
         if filename:
-            torch.save(checkpoint_model, filename)
+            torch.save(checkpoint_model, filename, pickle_module = dill)
         self.opt_loaded_state_dict = checkpoint_model["optimizer_state"]
         self.model_load_state_dict = checkpoint_model["model_state"]
         self.optimizer.load_state_dict(self.opt_loaded_state_dict)
@@ -822,6 +831,13 @@ class Trainer():
             num_workers (int, optional): How many subprocesses to use for data loading. 0 means that the data will be loaded in the main process.
                         Defaults to 0.
         """
[…]
