### 2023-10-19 a094e2f
add tests for testing when no test is provided and when no mmodel is loaded/trained
