### 2023-10-20 a6209cb
fix test_optim
