### 2023-10-20 d562513
change dataset_train to train_data and update docs, for later functionality refactor
