+++ deeprank2/dataset.py
@@ -30,6 +30,8 @@ class DeeprankDataset(Dataset):
     def __init__(self, # pylint: disable=too-many-arguments
                  hdf5_path: Union[str, List[str]],
                  subset: Optional[List[str]],
+                 train: bool,
+                 train_data: Optional[Union[str, GridDataset, GraphDataset]],
                  target: Optional[str],
                  task: Optional[str],
                  classes: Optional[Union[List[str], List[int], List[float]]],
@@ -54,11 +56,11 @@ class DeeprankDataset(Dataset):
         else:
             raise TypeError(f"hdf5_path: unexpected type: {type(hdf5_path)}")
 
-        self.use_tqdm = use_tqdm
-
+        self.train = train
+        self.train_data = train_data
         self.target = target
         self.subset = subset
-
+        self.use_tqdm = use_tqdm
         self.target_filter = target_filter
 
         if check_integrity:
@@ -412,7 +414,7 @@ class GridDataset(DeeprankDataset):
         hdf5_path: Union[str, list],
         subset: Optional[List[str]] = None,
         train: bool = True,
-        dataset_train: Optional[GridDataset] = None,
+        train_data: Optional[Union[str, GridDataset]] = None,
         features: Optional[Union[List[str], str]] = "all",
         target: Optional[str] = None,
         target_transform: Optional[bool] = False,
@@ -429,16 +431,16 @@ class GridDataset(DeeprankDataset):
             hdf5_path (Union[str,list]): Path to .HDF5 file(s). For multiple .HDF5 files, insert the paths in a List. Defaults to None.
             subset (Optional[List[str]], optional): List of keys from .HDF5 file to include. Defaults to None (meaning include all).
             train (bool, optional): Boolean flag to determine if the instance represents the training set.
-                If False, a dataset_train of the same class must be provided as well.
-                The latter will be used to scale the validation/testing set according to its features values and to match the datasets' parameters.
+                It needs to be set to False only if the instance represents the validation or testing set.
                 Defaults to True.
-            dataset_train (class:`GridDataset`, optional): If `train` is True, assign here the training set.
-                If `train` is False and `dataset_train` is assigned,
-                the parameters `features`, `target`, `traget_transform`, `task`, and `classes` will be inherited from `dataset_train`.
+            train_data (Optional[Union[str, class:`GridDataset`]], optional)): data to inherit information from the training dataset or the pre-trained model.
+                It is considered only if `train` is False. In the latter case, `train_data` needs to be a dataset of the same class or
+                the path of a DeepRank2 pre-trained model. If set, the parameters `features`, `target`, `traget_transform`, `task`, and `classes`
+                will be inherited from `train_data`.
                 Defaults to None.
             features (Optional[Union[List[str], str]], optional): Consider all pre-computed features ("all") or some defined node features
                 (provide a list, example: ["res_type", "polarity", "bsa"]). The complete list can be found in `deeprank2.domain.gridstorage`.
-                Value will be ignored and inherited from `dataset_train` if `train` is set as False and `dataset_train` is assigned.
+                Value will be ignored and inherited from `train_data` if `train` is set as False and `train_data` is assigned.
                 Defaults to "all".
             target (Optional[str], optional): Default options are irmsd, lrmsd, fnat, binary, capri_class, and dockq. It can also be
                 a custom-defined target given to the Query class as input (see: `deeprank2.query`); in this case,
@@ -446,11 +448,11 @@ class GridDataset(DeeprankDataset):
                 Only numerical target variables are supported, not categorical.
                 If the latter is your case, please convert the categorical classes into
                 numerical class indices before defining the :class:`GraphDataset` instance.
-                Value will be ignored and inherited from `dataset_train` if `train` is set as False and `dataset_train` is assigned.
[…]
