+++ tests/test_integration.py
@@ -82,13 +82,13 @@ def test_cnn(): # pylint: disable=too-many-locals
         dataset_val = GridDataset(
             hdf5_path = hdf5_paths,
             train = False,
-            dataset_train = dataset_train,
+            train_data = dataset_train,
         )
 
         dataset_test = GridDataset(
             hdf5_path = hdf5_paths,
             train = False,
-            dataset_train = dataset_train,
+            train_data = dataset_train,
         )
 
         output_exporters = [HDF5OutputExporter(output_directory)]
@@ -165,14 +165,14 @@ def test_gnn(): # pylint: disable=too-many-locals
         dataset_val = GraphDataset(
             hdf5_path = hdf5_paths,
             train = False,
-            dataset_train = dataset_train,
+            train_data = dataset_train,
             clustering_method = "mcl"
         )
 
         dataset_test = GraphDataset(
             hdf5_path = hdf5_paths,
             train = False,
-            dataset_train = dataset_train,
+            train_data = dataset_train,
             clustering_method = "mcl"
         )
 
