+++ README.md
@@ -216,13 +216,13 @@ dataset_val = GraphDataset(
     hdf5_path = hdf5_paths,
     subset = valid_ids,
     train = False,
-    dataset_train = dataset_train
+    train_data = dataset_train
 )
 dataset_test = GraphDataset(
     hdf5_path = hdf5_paths,
     subset = test_ids,
     train = False,
-    dataset_train = dataset_train
+    train_data = dataset_train
 )
 ```
 
@@ -250,13 +250,13 @@ dataset_val = GridDataset(
     hdf5_path = hdf5_paths,
     subset = valid_ids,
     train = False,
-    dataset_train = dataset_train,
+    train_data = dataset_train,
 )
 dataset_test = GridDataset(
     hdf5_path = hdf5_paths,
     subset = test_ids,
     train = False,
-    dataset_train = dataset_train,
+    train_data = dataset_train,
 )
 ```
 

[…]
