### 2023-10-20 fe18d3d
improve logic for handling both a pre-trained model and a dataset_train when train is False in dataset.py
