### 2023-10-21 f9b82de
add logic for handling the pre-trained model as input in DeeprankDataset classes
