### 2023-10-21 c21fe2b
add tests for catching uncorrect pre-trained models
