### 2023-10-23 bcc5138
implement inheritance in dataset from a pre-trained model
