### 2023-10-23 5280ece
add tests for inheritance from pre-trained model
