### 2023-10-23 7fcc033
add classes_to_index as inherited param and to the pre-trained model
