### 2023-10-23 6b037ad
add classes_to_index's check to the tests
