+++ tests/test_dataset.py
@@ -256,7 +256,7 @@ class TestDataSet(unittest.TestCase):
 
         # features, target, target_transform, task, and classes
         # in the test should be inherited from the pre-trained model
-        inherited_params = ["features", "target", "target_transform", "task", "classes"]
+        inherited_params = ["features", "target", "target_transform", "task", "classes", "classes_to_index"]
         data = torch.load(pretrained_model, pickle_module = dill, map_location=torch.device('cpu'))
 
         dataset_test_vars = vars(dataset_test)
@@ -1017,7 +1017,7 @@ class TestDataSet(unittest.TestCase):
 
         # node_features, edge_features, feature_transform, target, target_transform, task, and classes
         # in the test should be inherited from the pre-trained model
-        inherited_params = ["node_features", "edge_features", "features_transform", "target", "target_transform", "task", "classes"]
+        inherited_params = ["node_features", "edge_features", "features_transform", "target", "target_transform", "task", "classes", "classes_to_index"]
         data = torch.load(pretrained_model, pickle_module = dill, map_location=torch.device('cpu'))
 
         dataset_test_vars = vars(dataset_test)
