### 2023-10-23 0fcea3a
add trainer tests for testing without defining the dataset_train
