### 2023-10-23 b8e2348
remove dill usage since we're not saving lambda functions anymore (dill not needed)
