### 2023-10-23 060e6bf
improve initialization order in the Trainer class
