### 2023-10-24 4d588e1
fix Trainer _eval method for cases in which there is a target attribute but no target values are present in the hdf5 file/s
