### 2023-10-24 147a16a
add logic for checking the target settings in the init, and fix _filter_targets
