### 2023-10-24 08c90e0
add tests for cases with no target and improve target's filter tests
