### 2023-10-24 3b5d746
fix tests according to the new target's checks
