+++ tests/test_trainer.py
@@ -552,12 +552,14 @@ class TestTrainer(unittest.TestCase):
         dataset_val = GraphDataset(
             hdf5_path = "tests/data/hdf5/test.hdf5",
             train = True,
-            train_data = dataset_train
+            train_data = dataset_train,
+            target = targets.BINARY
         )
         dataset_test = GraphDataset(
             hdf5_path = "tests/data/hdf5/test.hdf5",
             train = True,
-            train_data = dataset_train
+            train_data = dataset_train,
+            target = targets.BINARY
         )
         with pytest.raises(ValueError):
             Trainer(
@@ -638,7 +640,7 @@ class TestTrainer(unittest.TestCase):
 
         for t in test_cases:
             dataset_train, dataset_val =_divide_dataset(
-                dataset = GraphDataset(hdf5_path = hdf5),
+                dataset = GraphDataset(hdf5_path = hdf5, target = targets.BINARY),
                 splitsize = t,
             )
             assert len(dataset_train) == n_train
@@ -668,7 +670,8 @@ class TestTrainer(unittest.TestCase):
 
     def test_invalid_cuda_ngpus(self):
         dataset_train = GraphDataset(
-            hdf5_path = "tests/data/hdf5/test.hdf5"
+            hdf5_path = "tests/data/hdf5/test.hdf5",
+            target = targets.BINARY
         )
         dataset_val = GraphDataset(
             hdf5_path = "tests/data/hdf5/test.hdf5",
@@ -687,7 +690,8 @@ class TestTrainer(unittest.TestCase):
     def test_invalid_no_cuda_available(self):
         if not torch.cuda.is_available():
             dataset_train = GraphDataset(
-                hdf5_path = "tests/data/hdf5/test.hdf5"
+                hdf5_path = "tests/data/hdf5/test.hdf5",
+                target = targets.BINARY
             )
             dataset_val = GraphDataset(
                 hdf5_path = "tests/data/hdf5/test.hdf5",
