### 2023-10-24 32f3e25
add test for verifying that the testing output is correct when target is not present
