### 2023-10-24 d842fc7
try to fix geometric installation using pip instead of conda
