+++ .github/actions/install-python-and-package/action.yml
@@ -51,7 +51,7 @@ runs:
         ## PyTorch, PyG, PyG adds
         ### Installing for CPU only on the CI
         conda install pytorch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 -c pytorch
-        conda install pyg -c pyg
+        pip install torch_geometric==2.3.1
         pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-$(python3 -c "import torch; print(torch.__version__)")+cpu.html
     - name: Install dependencies on MacOS
       shell: bash {0}
