### 2023-10-25 10a5795
add docs for testing a pre-trained model
