### 2023-10-25 #5983 @gcroci2
> @gcroci2 there's a few warnings about DOIs above, have you looked into these to see if they're meaningful?

They all are, the ones marked as invalid should just have the prefix 'https://doi.org/' removed. Should I do it myself in the .bib file in the repository?
About the missing ones, where should I insert them?
