### 2023-10-27 #5983 @abhishektiwari
@gcroci2 Given Mac and Windows are not supported, can  DeepRank2 run inside the latest Ubuntu container?

> The package officially supports ubuntu-latest OS only, whose functioning is widely tested through the continuous integration workflows.
