### 2023-10-30 #5983 @gcroci2
> @gcroci2 Given Mac and Windows are not supported, can DeepRank2 run inside the latest Ubuntu container?
>
> > The package officially supports ubuntu-latest OS only, whose functioning is widely tested through the continuous integration workflows.

As written in the README, yes it can.
