### 2023-10-30 bf1c8a3
fix bug in trainer for testing cases with no target
