### 2023-11-01 7f0a021
Update CONTRIBUTING.rst
