+++ pyproject.toml
@@ -49,7 +49,8 @@ dependencies = [
     "tqdm >= 4.63.0",
     "freesasa >= 2.1.0",
     "tensorboard >= 0.9.0",
-    "protobuf >= 3.20.1"
+    "protobuf >= 3.20.1",
+    "ruff"
 ]
 
 [project.optional-dependencies]
@@ -83,3 +84,31 @@ source = ["deeprank2"]
 [tool.setuptools.packages.find]
 include = ["deeprank2*"]
 exclude = ["tests*"]
+
+[tool.ruff]
+target-version = "py310"
+line-length = 159
+select = [
+    "F",    # Pyflakes
+    "E",    # pycodestyle (error)
+    "W",    # pycodestyle (warning)
+    "I",    # isort
+    "D",    # pydocstyle
+]
+ignore = [
+    # Disable because not part of PEP257 official convention:
+    # see http://pep257.readthedocs.io/en/latest/error_codes.html
+    "D203",  # 1 blank line required before class docstring
+    "D212",  # Multi-line docstring summary should start at the first line
+    "D213",  # Multi-line docstring summary should start at the second line
+    "D404",  # First word of the docstring should not be This
+]
+
+# Allow autofix for all enabled rules (when `--fix`) is provided.
+fixable = ["A", "B", "C", "D", "E", "F", "I"]
+
+[tool.ruff.isort]
+known-first-party = ["deeprank2"]
+force-single-line = true
+lines-after-imports = 2
+no-lines-before = ["future","standard-library","third-party","first-party","local-folder"]
