### 2023-11-07 9e426c5
format deeprank2 using ruff
