+++ deeprank2/utils/parsing/top.py
@@ -4,9 +4,9 @@ from typing import Any, Dict
 
 logging.getLogger(__name__)
 
+
 class TopRowObject:
-    def __init__(self, residue_name: str,
-                 atom_name: str, kwargs: Dict[str, Any]):
+    def __init__(self, residue_name: str, atom_name: str, kwargs: Dict[str, Any]):
         self.residue_name = residue_name
         self.atom_name = atom_name
         self.kwargs = kwargs
@@ -14,11 +14,10 @@ class TopRowObject:
     def __getitem__(self, key):
         return self.kwargs[key]
 
+
 class TopParser:
     _VAR_PATTERN = re.compile(r"([^\s]+)\s*=\s*([^\s\(\)]+|\(.*\))")
-    _LINE_PATTERN = re.compile(
-        r"^([A-Z0-9]{3})\s+atom\s+([A-Z0-9]{1,4})\s+(.+)\s+end\s*(\s+\!\s+[ _A-Za-z0-9]+)?$"
-    )
+    _LINE_PATTERN = re.compile(r"^([A-Z0-9]{3})\s+atom\s+([A-Z0-9]{1,4})\s+(.+)\s+end\s*(\s+\!\s+[ _A-Za-z0-9]+)?$")
     _NUMBER_PATTERN = re.compile(r"\-?[0-9]+(\.[0-9]+)?")
 
     @staticmethod
@@ -35,9 +34,7 @@ class TopParser:
 
             kwargs = {}
             for w in TopParser._VAR_PATTERN.finditer(m.group(3)):
-                kwargs[w.group(1).lower().strip()] = TopParser._parse_value(
-                    w.group(2).strip()
-                )
+                kwargs[w.group(1).lower().strip()] = TopParser._parse_value(w.group(2).strip())
 
             result.append(TopRowObject(residue_name, atom_name, kwargs))
 
