+++ deeprank2/utils/graph.py
@@ -23,9 +23,7 @@ class Edge:
         self.id = id_
         self.features = {}
 
-    def add_feature(
-        self, feature_name: str, feature_function: Callable[[Contact], float]
-    ):
+    def add_feature(self, feature_name: str, feature_function: Callable[[Contact], float]):
         feature_value = feature_function(self.id)
 
         self.features[feature_name] = feature_value
@@ -82,9 +80,7 @@ class Node:
 
         if len(feature_value.shape) != 1:
             shape_s = "x".join(feature_value.shape)
-            raise ValueError(
-                f"Expected a 1-dimensional array for feature {feature_name}, but got {shape_s}"
-            )
+            raise ValueError(f"Expected a 1-dimensional array for feature {feature_name}, but got {shape_s}")
 
         self.features[feature_name] = feature_value
 
@@ -140,18 +136,19 @@ class Graph:
 
         return False
 
-    def _map_point_features(self, grid: Grid, method: MapMethod,  # pylint: disable=too-many-arguments
-                            feature_name: str, points: List[np.ndarray],
-                            values: List[Union[float, np.ndarray]],
-                            augmentation: Optional[Augmentation] = None):
-
+    def _map_point_features(
+        self,
+        grid: Grid,
+        method: MapMethod,  # pylint: disable=too-many-arguments
+        feature_name: str,
+        points: List[np.ndarray],
+        values: List[Union[float, np.ndarray]],
+        augmentation: Optional[Augmentation] = None,
+    ):
         points = np.stack(points, axis=0)
 
         if augmentation is not None:
-            points = pdb2sql.transform.rot_xyz_around_axis(points,
-                                                           augmentation.axis,
-                                                           augmentation.angle,
-                                                           self.center)
+            points = pdb2sql.transform.rot_xyz_around_axis(points, augmentation.axis, augmentation.angle, self.center)
 
         for point_index in range(points.shape[0]):
             position = points[point_index]
@@ -160,12 +157,10 @@ class Graph:
             grid.map_feature(position, feature_name, value, method)
 
     def map_to_grid(self, grid: Grid, method: MapMethod, augmentation: Optional[Augmentation] = None):
-
         # order edge features by xyz point
         points = []
         feature_values = {}
         for edge in self._edges.values():
-
             points += [edge.position1, edge.position2]
 
             for feature_name, feature_value in edge.features.items():
@@ -179,7 +174,6 @@ class Graph:
         points = []
         feature_values = {}
         for node in self._nodes.values():
-
             points.append(node.position)
 
             for feature_name, feature_value in node.features.items():
@@ -189,11 +183,10 @@ class Graph:
         for feature_name, values in feature_values.items():
             self._map_point_features(grid, method, feature_name, points, values, augmentation)
 
-    def write_to_hdf5(self, hdf5_path: str): # pylint: disable=too-many-locals
+    def write_to_hdf5(self, hdf5_path: str):  # pylint: disable=too-many-locals
         """Write a featured graph to an hdf5 file, according to deeprank standards."""
 
         with h5py.File(hdf5_path, "a") as hdf5_file:
-
             # create groups to hold data
             graph_group = hdf5_file.require_group(self.id)
             node_features_group = graph_group.create_group(Nfeat.NODE)
@@ -210,14 +203,9 @@ class Graph:
             first_node_data = list(self._nodes.values())[0].features
             node_feature_names = list(first_node_data.keys())
             for node_feature_name in node_feature_names:
[…]
