+++ tests/utils/test_community_pooling.py
@@ -2,17 +2,13 @@ import unittest
 
 import numpy as np
 import torch
-from deeprank2.utils.community_pooling import (community_detection,
-                                               community_detection_per_batch,
-                                               community_pooling)
+from deeprank2.utils.community_pooling import community_detection, community_detection_per_batch, community_pooling
 from torch_geometric.data import Batch, Data
 
 
 class TestCommunity(unittest.TestCase):
     def setUp(self):
-        self.edge_index = torch.tensor(
-            [[0, 1, 1, 2, 3, 4, 4, 5], [1, 0, 2, 1, 4, 3, 5, 4]], dtype=torch.long
-        )
+        self.edge_index = torch.tensor([[0, 1, 1, 2, 3, 4, 4, 5], [1, 0, 2, 1, 4, 3, 5, 4]], dtype=torch.long)
 
         self.x = torch.tensor([[0], [1], [2], [3], [4], [5]], dtype=torch.float)
         self.data = Data(x=self.x, edge_index=self.edge_index)
@@ -22,9 +18,7 @@ class TestCommunity(unittest.TestCase):
         community_detection(self.data.edge_index, self.data.num_nodes, method="mcl")
 
     def test_detection_louvain(self):
-        community_detection(
-            self.data.edge_index, self.data.num_nodes, method="louvain"
-        )
+        community_detection(self.data.edge_index, self.data.num_nodes, method="louvain")
 
     @unittest.expectedFailure
     def test_detection_error(self):
@@ -59,7 +53,6 @@ class TestCommunity(unittest.TestCase):
         )
 
     def test_pooling(self):
-
         batch = Batch().from_data_list([self.data, self.data])
         cluster = community_detection(batch.edge_index, batch.num_nodes)
 

[…]
