+++ tests/domain/test_aminoacidlist.py
@@ -1,5 +1,10 @@
 import numpy as np
-from deeprank2.domain.aminoacidlist import amino_acids, cysteine, lysine, pyrrolysine, selenocysteine
+from deeprank2.domain.aminoacidlist import amino_acids
+from deeprank2.domain.aminoacidlist import cysteine
+from deeprank2.domain.aminoacidlist import lysine
+from deeprank2.domain.aminoacidlist import pyrrolysine
+from deeprank2.domain.aminoacidlist import selenocysteine
+
 
 # Exceptions selenocysteine and pyrrolysine are due to them having the same index as their canonical counterpart.
 # This is not an issue while selenocysteine and pyrrolysine are not part of amino_acids.
