+++ tests/utils/test_grid.py
@@ -1,7 +1,10 @@
 import h5py
 import numpy as np
-from deeprank2.query import ProteinProteinInterfaceAtomicQuery, ProteinProteinInterfaceResidueQuery
-from deeprank2.utils.grid import Grid, GridSettings, MapMethod
+from deeprank2.query import ProteinProteinInterfaceAtomicQuery
+from deeprank2.query import ProteinProteinInterfaceResidueQuery
+from deeprank2.utils.grid import Grid
+from deeprank2.utils.grid import GridSettings
+from deeprank2.utils.grid import MapMethod
 
 
 def test_residue_grid_orientation():
