+++ tests/utils/test_earlystopping.py
@@ -1,5 +1,6 @@
 from deeprank2.utils.earlystopping import EarlyStopping
 
+
 dummy_val_losses = [3, 2, 1, 2, 0.5, 2, 3, 4, 5, 6, 7]
 dummy_train_losses = [3, 2, 1, 2, 0.5, 2, 3, 4, 5, 1, 7]
 
