+++ tests/utils/test_community_pooling.py
@@ -1,9 +1,11 @@
 import unittest
-
 import numpy as np
 import torch
-from deeprank2.utils.community_pooling import community_detection, community_detection_per_batch, community_pooling
-from torch_geometric.data import Batch, Data
+from torch_geometric.data import Batch
+from torch_geometric.data import Data
+from deeprank2.utils.community_pooling import community_detection
+from deeprank2.utils.community_pooling import community_detection_per_batch
+from deeprank2.utils.community_pooling import community_pooling
 
 
 class TestCommunity(unittest.TestCase):
