+++ tests/features/test_contact.py
@@ -112,7 +112,7 @@ def test_14_pair():
 def test_14dist_opposing_chains():
     """ChainA PRO 114 CA - ChainB HIS 116 CD2 (3.62 A). Should have non-zero energy despite close contact, because opposing chains.
     E_vdw for this pair if they were on the same chain: 0.018
-    E_vdw for this pair on opposing chains: 0.146
+    E_vdw for this pair on opposing chains: 0.146.
     """
     opposing_edge = _get_contact("1A0Z", 114, "CA", 116, "CD2", chains=("A", "B"))
     assert opposing_edge.features[Efeat.DISTANCE] > cutoff_13
