### 2023-11-07 9e8b228
remove prospector refs
