+++ pyproject.toml
@@ -57,8 +57,6 @@ dependencies = [
 # development dependency groups
 test = [
     "pytest >= 7.4.0",
-    "pylint <= 2.17.5",
-    "prospector[with_pyroma] <= 1.10.2",
     "bump2version",
     "coverage",
     "pycodestyle",
