+++ deeprank2/features/conservation.py
@@ -8,9 +8,7 @@ from deeprank2.molstruct.residue import SingleResidueVariant
 from deeprank2.utils.graph import Graph
 
 
-def add_features(  # pylint: disable=unused-argument
-    pdb_path: str, graph: Graph, single_amino_acid_variant: Optional[SingleResidueVariant] = None
-):
+def add_features(pdb_path: str, graph: Graph, single_amino_acid_variant: Optional[SingleResidueVariant] = None):
     profile_amino_acid_order = sorted(amino_acids, key=lambda aa: aa.three_letter_code)
 
     for node in graph.nodes:
