+++ tests/domain/test_forcefield.py
@@ -10,7 +10,7 @@ def test_atomic_forcefield():
     try:
         structure = get_structure(pdb, "101M")
     finally:
-        pdb._close()  # pylint: disable=protected-access
+        pdb._close()
 
     # The arginine C-zeta should get a positive charge
     arg = [r for r in structure.get_chain("A").residues if r.amino_acid == arginine][0]
