### 2023-11-07 #5983 @jnwei
Installation: The installation instructions seem correct, but the steps for some of the packages (e.g. dssp) are somewhat cumbersome on a Mac or on a linux machine without sudo accesses. I did not test the installation procedure completely.

@gcroci2 To make installation easier for users, is it possible to provide a Docker container or Dockerfile for the repository?
