### 2023-11-08 0bbe2aa
fix training uml link
