### 2023-11-08 d1716ee
fix typo in training uml link
