### 2023-11-08 0d07c18
fix link for training uml
