+++ README.dev.md
@@ -80,4 +80,4 @@ During the development cycle, three main supporting branches are used:
 
 Code-base class diagrams updated on 02/11/2023, generated with https://www.gituml.com:
 - [Data processing classes and functions](https://gituml-media.s3.ap-southeast-2.amazonaws.com/production_diagram_2026.svg?AWSAccessKeyId=AKIA5BNPSF2PVKDZ4QNO&Signature=1AfRXogJj5JwWqIehv8vzdUC5So%3D&Expires=1698939357)
-- [ML pipeline classes and functions](https://gituml-media.s3.ap-southeast-2.amazonaws.com/production_diagram_2025.svg?AWSAccessKeyId=AKIA5BNPSF2PVKDZ4QNO&Signature=eeYNfbm8yYfYfkVgRJyRhv9D1zs%3D&Expires=1698939046](https://plantuml.atug.com/svg/xLhTRziu4dyNiF-3WNlPFLoCsaLVYY90jacp1T9SaMHnZmGZqJ8HYjIIb1FlNl_t6tv9f2JB7vlUmztwnUHmXfmP_ZWSaiePLcg64ELK8uJgxUp3XqeK3IFdvr2S_VZb7qBcQrhyD0F2townVBed-5eImcT_P3Fu8RGXKb714Sg-PRdjxvAGMcB-T8arLaHxJjj6L5n6Mi3G1Ncnn89iQ4v2qJ4MA_ktBbRlK8trsb6j329SI-exh0ZchgIQHnZJ5ZMMPL94AyeqaPwYdbn1Oigx4Yf0Qrjj547wjwBoTZeV_VYJg-PhaZyXgnTqIbHAftnD9DMa0AKbhdgjPcY4UO6y6zfMKTMD9iWOWvGkMhea6CY9rQ7DMOISVO5grGBVrZ083aH6k2yfl25eJPM6-hSi2zFquTzd_QdB-bFvzxEpyvasc9YxwPgz3_FMPvtFpsVhT_LsVYr5KmEVqNoBIbFBF5Ak9IrIb1HaXHkcqGPBDvyhiBsH9AsWWkR9nCBKSRKIScgoWh2dWbZmZC054fo0RhJmFkaPOtmJATlfZ-jrQW3UOChm8oE852L1YVfT6nT5hordZOAPfBo4-TThKOoxerqPmzw7pcafgNkwnJ9KuvNz2UwPScZdFU_GmjLEYKJNTvylqTUBkuVF5pVeqzN5mwztL-ZsuiiL41yUhkvkdUvCbBPmTdOMBGQpO9x8LZbb8otcyq4uUp_B1TSGAjJilCSQutcIBOBDBZwpqa3P4X0n1z_UG958u6vC4EtnIJ8tmGC2qGu6kn2lNW1W7BCRofy2N_wdTe1lhoJ_aVCxKWAMp4uIzWl1Dmnl8RYzUVDcG5aCACi1vQrR17dQIrTVZ88RKF-CPT67DpBBmlIEHACXJ7kovtQuSO6Hy_aNwm4ByiIqB3KqYgOWz14wVriuF4IMJubymIzVXM03
[…]
