### 2023-11-08 #5983 @gcroci2
> Installation: The installation instructions seem correct, but the steps for some of the packages (e.g. dssp) are somewhat cumbersome on a Mac or on a linux machine without sudo accesses. I did not test the installation procedure completely.
>
> @gcroci2 To make installation easier for users, is it possible to provide a Docker container or Dockerfile for the repository?

Officially we don't support Mac, but Linux only. We do provide references for DSSP installation on Mac though, and same for Linux machines without sudo accesses (https://github.com/pdb-redo/dssp). We may provide a Docker container for the repository, but we usually prefer to avoid that if it's not strictly necessary, as (in my opinion) in this case. Please let me know in case having the Docker container is a "mandatory" request for the paper to go through.
