### 2023-11-08 d050167
replace prospector with ruff
