+++ pyproject.toml
@@ -49,15 +49,14 @@ dependencies = [
     "tqdm >= 4.63.0",
     "freesasa >= 2.1.0",
     "tensorboard >= 0.9.0",
-    "protobuf >= 3.20.1"
+    "protobuf >= 3.20.1",
+    "ruff"
 ]
 
 [project.optional-dependencies]
 # development dependency groups
 test = [
     "pytest >= 7.4.0",
-    "pylint <= 2.17.5",
-    "prospector[with_pyroma] <= 1.10.2",
     "bump2version",
     "coverage",
     "pycodestyle",
@@ -83,3 +82,51 @@ source = ["deeprank2"]
 [tool.setuptools.packages.find]
 include = ["deeprank2*"]
 exclude = ["tests*"]
+
+[tool.ruff]
+# Exclude a variety of commonly ignored directories.
+exclude = [
+    "docs",
+    "reduce"
+]
+target-version = "py310"
+line-length = 159
+select = [
+    "F",    # Pyflakes
+    "E",    # pycodestyle (error)
+    "W",    # pycodestyle (warning)
+    "I",    # isort
+    "D",    # pydocstyle
+]
+ignore = [
+    "D100",  # Missing module docstring
+    "D101",  # Missing docstring in public class
+    "D102",  # Missing docstring in public method
+    "D103",  # Missing docstring in public function
+    "D104",  # Missing public package docstring
+    "D105",  # Missing docstring in magic method
+    "D107",  # Missing docstring in `__init__`
+    # The following list excludes rules irrelevant to the Google style
+    "D203",  # 1 blank line required before class docstring
+    "D204",
+    "D212",  # Multi-line docstring summary should start at the first line
+    "D213",  # Multi-line docstring summary should start at the second line
+    "D215",
+    "D400",
+    "D401",
+    "D404",  # First word of the docstring should not be This
+    "D406",
+    "D407",
+    "D408",
+    "D409",
+    "D413",
+]
+
+# Allow autofix for all enabled rules (when `--fix`) is provided.
+fixable = ["A", "B", "C", "D", "E", "F", "I"]
+
+[tool.ruff.isort]
+known-first-party = ["deeprank2"]
+force-single-line = true
+lines-after-imports = 2
+no-lines-before = ["future","standard-library","third-party","first-party","local-folder"]
