### 2023-11-09 4323410
upload svgs and try to link them to the readme.dev
