### 2023-11-09 cd2f502
improve svg image visualization for umls
