+++ README.dev.md
@@ -78,7 +78,6 @@ During the development cycle, three main supporting branches are used:
 
 ## UML
 
-Code-base class diagrams updated on 02/11/2023, generated with https://www.gituml.com (click on the image for zooming):
-- Data processing classes and functions
-  <img src="./tests/utils/uml_data_processing.svg" width="50">
-- [ML pipeline classes and functions]()
+Code-base class diagrams updated on 02/11/2023, generated with https://www.gituml.com (save the images and open them in the browser for zooming).
+- Data processing classes and functions: <img src="./tests/utils/uml_data_processing.svg" width="50">
+- ML pipeline classes and functions: <img src="./tests/utils/uml_training.svg" width="50">
