### 2023-11-09 #5983 @jnwei
I found instructions for DSSP installation to be quite non trivial, requiring several C++ packages to be built from source code on git. I was unable to complete installation, and thus I was not able to verify that the tests pass or that the code run. Perhaps another reviewer who does have a Linux machine could quickly check that the tests and notebooks run as intended.

The dockerfile could help potential users explore your package quickly with minimal overhead. However, I will leave the decision of whether to include it up to you.

I instead focused my review on reading the documentation and the tutorial notebooks. Overall I think the documentation is high quality. The code looks correct upon visual inspection, and accounts properly for random seeds.

A few small suggestions:
- I noticed a few potential typos, and suggested fixes for them here: https://github.com/DeepRank/deeprank2/pull/523
- It could be helpful to provide an environment.yml file to support the python dependencies (e.g. pytorch, pytorch geometric + optional packages).
