### 2023-11-11 #5983 @abhishektiwari
I second comments of @jnwei on `installation instructions`.

@gcroci2 I tried `Ubuntu` Docker image, end-up creating a `Dockerfile` to run the package on Mac, but I am not sure if I have all components. We will find out soon. It will be better if an official `Dockerfile` was provided for the package. I think it will reduce friction anyone looking to use this package or contribute.
