### 2023-11-11 #5983 @abhishektiwari
@gcroci2 Please see my [bug report](https://github.com/DeepRank/deeprank2/issues/524) when running with `ubuntu:latest` Docker container. Added custom `Dockerfile` I created to run `deeprank2` on Mac with instructions to get notebook up and running.
