### 2023-11-14 edab689
fix installation bug
