### 2023-11-14 #5983 @gcroci2
> * I noticed a few potential typos, and suggested fixes for them here: [Joss review 5983 jnwei DeepRank/deeprank2#523](https://github.com/DeepRank/deeprank2/pull/523)

Thanks, I just merged it @jnwei
