+++ tests/utils/test_grid.py
@@ -1,11 +1,12 @@
 import h5py
 import numpy as np
+
 from deeprank2.query import (ProteinProteinInterfaceAtomicQuery,
                              ProteinProteinInterfaceResidueQuery)
 from deeprank2.utils.grid import Grid, GridSettings, MapMethod
 
 
-def test_residue_grid_orientation():
+def test_residue_grid_orientation(): # pylint: disable=too-many-locals
 
     coord_error_margin = 1.0  # Angstrom
 
@@ -53,7 +54,7 @@ def test_residue_grid_orientation():
     assert np.all(np.abs(grid.zs - target_zs) < coord_error_margin), f"\n{grid.zs} != \n{target_zs}"
 
 
-def test_atomic_grid_orientation():
+def test_atomic_grid_orientation(): # pylint: disable=too-many-locals
 
     coord_error_margin = 1.0  # Angstrom
 
@@ -96,6 +97,9 @@ def test_atomic_grid_orientation():
 
     assert grid.ys.shape == target_ys.shape
     assert np.all(np.abs(grid.ys - target_ys) < coord_error_margin), f"\n{grid.ys} != \n{target_ys}"
+
+    assert grid.zs.shape == target_zs.shape
+    assert np.all(np.abs(grid.zs - target_zs) < coord_error_margin), f"\n{grid.zs} != \n{target_zs}"
 
     assert grid.zs.shape == target_zs.shape
     assert np.all(np.abs(grid.zs - target_zs) < coord_error_margin), f"\n{grid.zs} != \n{target_zs}"
