+++ tests/molstruct/test_structure.py
@@ -2,9 +2,10 @@
 import pickle
 from multiprocessing.connection import _ForkingPickler
 
+from pdb2sql import pdb2sql
+
 from deeprank2.molstruct.structure import PDBStructure
 from deeprank2.utils.buildgraph import get_structure
-from pdb2sql import pdb2sql
 
 
 def _get_structure(path) -> PDBStructure:
@@ -12,7 +13,7 @@ def _get_structure(path) -> PDBStructure:
     try:
         structure = get_structure(pdb, "101M")
     finally:
-        pdb._close()
+        pdb._close() # pylint: disable=protected-access
 
     assert structure is not None
 
@@ -44,4 +45,5 @@ def test_serialization_fork():
     assert loaded_structure.get_chain("A") == structure.get_chain("A")
     assert loaded_structure.get_chain("A").get_residue(0) == structure.get_chain("A").get_residue(0)
     assert loaded_structure.get_chain("A").get_residue(0).amino_acid == structure.get_chain("A").get_residue(0).amino_acid
+    assert loaded_structure.get_chain("A").get_residue(0).atoms[0] == structure.get_chain("A").get_residue(0).atoms[0]
     assert loaded_structure.get_chain("A").get_residue(0).atoms[0] == structure.get_chain("A").get_residue(0).atoms[0]
