### 2023-11-14 #5983 @gcroci2
> * It could be helpful to provide an environment.yml file to support the python dependencies (e.g. pytorch, pytorch geometric + optional packages).

We can do that, but which pytorch version should I use then? There are different ways of installing pytorch depending on the cuda version. I could just go for the same version we use for the CI (cuda 11.7), but then it's not going to work if the user has cpu only.

Also, the pytorch geometric extra dependencies (torch_scatter, torch_sparse, torch_cluster, torch_spline_conv) are installed via pip in the following way: `pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-${TORCH_VERSION}+cpu.html`, so I am not sure about how to include the extra link in the yml file.

Any suggestions is welcome @jnwei
