+++ tests/utils/test_grid.py
@@ -1,8 +1,10 @@
 import h5py
 import numpy as np
 
-from deeprank2.query import (ProteinProteinInterfaceAtomicQuery,
-                             ProteinProteinInterfaceResidueQuery)
+from deeprank2.query import (
+    ProteinProteinInterfaceAtomicQuery,
+    ProteinProteinInterfaceResidueQuery,
+)
 from deeprank2.utils.grid import Grid, GridSettings, MapMethod
 
 
@@ -97,9 +99,6 @@ def test_atomic_grid_orientation(): # pylint: disable=too-many-locals
 
     assert grid.ys.shape == target_ys.shape
     assert np.all(np.abs(grid.ys - target_ys) < coord_error_margin), f"\n{grid.ys} != \n{target_ys}"
-
-    assert grid.zs.shape == target_zs.shape
-    assert np.all(np.abs(grid.zs - target_zs) < coord_error_margin), f"\n{grid.zs} != \n{target_zs}"
 
     assert grid.zs.shape == target_zs.shape
     assert np.all(np.abs(grid.zs - target_zs) < coord_error_margin), f"\n{grid.zs} != \n{target_zs}"
