+++ CITATION.cff
@@ -55,5 +55,5 @@ keywords:
   - DeepRank
 license: Apache-2.0
 commit: 4e8823758ba03f824b4281f5689cb6a335ab2f6c
-version: "2.1.0"
+version: "2.1.1"
 date-released: '2023-09-22'
