### 2023-11-17 #5983 @gcroci2
> @gcroci2 Please see my [bug report](https://github.com/DeepRank/deeprank2/issues/524) when running with `ubuntu:latest` Docker container. Added custom `Dockerfile` I created to run `deeprank2` on Mac with instructions to get notebook up and running.

Hi @abhishektiwari, thanks for the report. Now the bug should be fixed and we released [a new version of the software](https://pypi.org/project/deeprank2/2.1.1/). Feel free to try it out again.

@jnwei We will work on creating an official Dockerfile (and likely an environment.yml as you suggested) in the coming days, and we will try to have it ready by the end of the next week.
