### 2023-11-17 d5acd2c
update `max_edge_length` and `influence_radius`
previously `max_edge_distance` and `interaction_radius`, respectively
