+++ tests/features/test_irc.py
@@ -28,8 +28,8 @@ def test_irc_residue():
     graph, _ = build_testgraph(
         pdb_path=pdb_path,
         detail='residue',
-        interaction_radius=8.5,
-        max_edge_distance=8.5,
+        influence_radius=8.5,
+        max_edge_length=8.5,
     )
     add_features(pdb_path, graph)
     _run_assertions(graph)
@@ -40,8 +40,8 @@ def test_irc_atom():
     graph, _ = build_testgraph(
         pdb_path=pdb_path,
         detail='residue',
-        interaction_radius=4.5,
-        max_edge_distance=4.5,
+        influence_radius=4.5,
+        max_edge_length=4.5,
     )
     add_features(pdb_path, graph)
     _run_assertions(graph)
