### 2023-11-17 86b15fb
put redundant code for inheriting training info in the parent class
