+++ tests/test_query.py
@@ -47,31 +47,31 @@ def _check_graph_makes_sense(
         g.write_to_hdf5(tmp_path)
 
         with h5py.File(tmp_path, "r") as f5:
-            entry_group = f5[list(f5.keys())[0]]
+            grp = f5[list(f5.keys())[0]]
             for feature_name in node_feature_names:
                 assert (
-                    entry_group[f"{Nfeat.NODE}/{feature_name}"][()].size > 0
+                    grp[f"{Nfeat.NODE}/{feature_name}"][()].size > 0
                 ), f"no {feature_name} feature"
 
                 assert (
                     len(
                         np.nonzero(
-                            entry_group[f"{Nfeat.NODE}/{feature_name}"][()]
+                            grp[f"{Nfeat.NODE}/{feature_name}"][()]
                         )
                     )
                     > 0
                 ), f"{feature_name}: all zero"
 
-            assert entry_group[f"{Efeat.EDGE}/{Efeat.INDEX}"][()].shape[1] == 2, "wrong edge index shape"
-            assert entry_group[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0] > 0, "no edge indices"
+            assert grp[f"{Efeat.EDGE}/{Efeat.INDEX}"][()].shape[1] == 2, "wrong edge index shape"
+            assert grp[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0] > 0, "no edge indices"
 
             for feature_name in edge_feature_names:
                 assert (
-                    entry_group[f"{Efeat.EDGE}/{feature_name}"][()].shape[0]
-                    == entry_group[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0]
+                    grp[f"{Efeat.EDGE}/{feature_name}"][()].shape[0]
+                    == grp[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0]
                 ), f"not enough edge {feature_name} feature values"
 
-            count_edges_hdf5 = entry_group[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0]
+            count_edges_hdf5 = grp[f"{Efeat.EDGE}/{Efeat.INDEX}"].shape[0]
 
         dataset = GraphDataset(hdf5_path=tmp_path, target=targets.BINARY)
         torch_data_entry = dataset[0]
