### 2023-11-17 6a6a305
make inherit_params an attribute of the dataset classes
