### 2023-11-21 55c061c
make error msg about the dataset clearer
