### 2023-11-21 2584917
use None instead of 'None' in the trainer _eval and _epoch methods
