### 2023-11-21 0906b2d
fix prospector errors
