### 2023-11-22 836c5b2
move features checking after inheritance
