### 2023-11-22 780f2b9
try to fix optimizer error in py3.11
