### 2023-11-24 2935414
Update docs/getstarted.md
Co-authored-by: Dani Bodor <d.bodor@esciencecenter.nl>
