### 2023-11-27 ef80c23
add files for building conda env
