+++ env/requirements.txt
@@ -1,0 +1,5 @@
+--find-links https://data.pyg.org/whl/torch-2.1.0+cpu.html
+torch_scatter==2.1.2
+torch_sparse==0.6.18
+torch_cluster==1.6.3
+torch_spline_conv==1.2.2
