### 2023-11-27 d3cf4f9
finalize dockerfile
