### 2023-11-27 f8d1a10
update docs for running docker and installation using yml file
