### 2023-11-28 cc6f5ea
docs: fix TOC alignment in README
Somehow the alignment of the table of contents got messed up. This PR fixes that
