+++ docs/installation.md
@@ -1,33 +1,36 @@
-# Installation
+# Installations
 
 Note that the package officially supports ubuntu-latest OS only, whose functioning is widely tested through the continuous integration workflows. 
 
-## Dockerfile
+You can either install DeepRank2 in a [dockerized container](#containerized-installation), which will allow you to run our [tutorial notebooks](https://github.com/DeepRank/deeprank2/tree/main/tutorials), or you can [install the package locally](#localremote-installation).
 
-In order to try out the package without worrying about your OS and without the need of installing all the required dependencies, we created a `Dockerfile` that can be used for taking care of everything in a suitable container. After having cloned the repository and installed [Docker](https://docs.docker.com/engine/install/), run the following commands from the root of the repository.
+## Containerized Installation 
+
+In order to try out the package without worrying about your OS and without the need of installing all the required dependencies, we created a `Dockerfile` that can be used for taking care of everything in a suitable container. After having cloned the repository and installed [Docker](https://docs.docker.com/engine/install/), run the following commands (you may need to have sudo permission) from the root of the repository.
 
 Build the Docker image:
+
 ```bash
 docker build -t deeprank2 .
 ```
 
-SSH to a running container:
+Run the Docker container:
 
 ```bash
-docker run -it --expose 3000  -p 3000:3000 deeprank2
+docker run -p 8888:8888 deeprank2
 ```
 
-Run the tutorials' notebooks from within the running container:
-```bash
-cd tutorials
-jupyter notebook --ip 0.0.0.0 --no-browser --allow-root --port 3000
-```
+This assumes that your application inside the container is listening on port 8888, and you want to map it to port 8888 on your host machine. Open a browser and go to `http://localhost:8888` to access the application running inside the Docker container and run the tutorials' notebooks.
 
-Now you can run the tutorials' notebook. More details about their content can be found [here](https://github.com/DeepRank/deeprank2/blob/main/tutorials/TUTORIAL.md). Note that in the docker container only the raw PDB files are downloaded, needed as a starting point for the tutorials. You can obtain the processed HDF5 files by running the `data_generation_xxx.ipynb` notebooks. Because Docker containers are limited in memory resources, we limit the number of data points processed in the tutorials'. Please install the package locally to fully leverage its capabilities.
+More details about the tutorials' content can be found [here](https://github.com/DeepRank/deeprank2/blob/main/tutorials/TUTORIAL.md). Note that in the docker container only the raw PDB files are downloaded, needed as a starting point for the tutorials. You can obtain the processed HDF5 files by running the `data_generation_xxx.ipynb` notebooks. Because Docker containers are limited in memory resources, we limit the number of data points processed in the tutorials'. Please install the package locally to fully leverage its capabilities.
 
-## Non-pythonic dependencies
+After running the tutorials, you may want to remove the (quite large) Docker image from your machine. In this case, remember to [stop the container](https://docs.docker.com/engine/reference/commandline/stop/) and then [remove the image](https://docs.docker.com/engine/reference/commandline/image_rm/). More general information about Docker can be found on the [official website docs](https://docs.docker.com/get-started/).
 
-Instructions are updated as of 14/09/2023.
+## Local/remote installation
+
+### Non-pythonic dependencies
+
+Instructions are up to date as of 14/09/2023.
 
 Before installing deeprank2 you need to install some dependencies:
 
@@ -38,26 +41,31 @@ Before installing deeprank2 you need to install some dependencies:
 *  [GCC](https://gcc.gnu.org/install/)
     * Check if gcc is installed: `gcc --version`. If this gives an error, run `sudo apt-get install gcc`. 
 
-## Pythonic dependencies
[…]
