+++ README.md
@@ -36,12 +36,14 @@ DeepRank2 extensive documentation can be found [here](https://deeprank2.rtfd.io/
 - [Deeprank2](#deeprank2)
   - [Overview](#overview)
   - [Table of contents](#table-of-contents)
-  - [Installation](#installation)
-    - [Dockerfile](#dockerfile)
-    - [Non-pythonic dependencies](#non-pythonic-dependencies)
-    - [Pythonic dependencies](#pythonic-dependencies)
+  - [Installations](#installations)
+    - [Containerized Installation](#containerized-installation)
+    - [Local/remote installation](#localremote-installation)
+      - [Non-pythonic dependencies](#non-pythonic-dependencies)
+      - [Pythonic dependencies](#pythonic-dependencies)
+      - [Install DeepRank2](#install-deeprank2)
       - [Test installation](#test-installation)
-    - [Contributing](#contributing)
+  - [Contributing](#contributing)
   - [Data generation](#data-generation)
   - [Datasets](#datasets)
     - [GraphDataset](#graphdataset)
@@ -50,36 +52,39 @@ DeepRank2 extensive documentation can be found [here](https://deeprank2.rtfd.io/
   - [Computational performances](#computational-performances)
   - [Package development](#package-development)
 
-## Installation
+## Installations
 
 Note that the package officially supports ubuntu-latest OS only, whose functioning is widely tested through the continuous integration workflows. 
 
-### Dockerfile
+You can either install DeepRank2 in a [dockerized container](#containerized-installation), which will allow you to run our [tutorial notebooks](https://github.com/DeepRank/deeprank2/tree/main/tutorials), or you can [install the package locally](#localremote-installation).
 
-In order to try out the package without worrying about your OS and without the need of installing all the required dependencies, we created a `Dockerfile` that can be used for taking care of everything in a suitable container. After having cloned the repository and installed [Docker](https://docs.docker.com/engine/install/), run the following commands from the root of the repository.
+### Containerized Installation 
+
+In order to try out the package without worrying about your OS and without the need of installing all the required dependencies, we created a `Dockerfile` that can be used for taking care of everything in a suitable container. After having cloned the repository and installed [Docker](https://docs.docker.com/engine/install/), run the following commands (you may need to have sudo permission) from the root of the repository.
 
 Build the Docker image:
+
 ```bash
 docker build -t deeprank2 .
 ```
 
-SSH to a running container:
+Run the Docker container:
 
 ```bash
-docker run -it --expose 3000  -p 3000:3000 deeprank2
+docker run -p 8888:8888 deeprank2
 ```
 
-Run the tutorials' notebooks from within the running container:
-```bash
-cd tutorials
-jupyter notebook --ip 0.0.0.0 --no-browser --allow-root --port 3000
-```
+This assumes that your application inside the container is listening on port 8888, and you want to map it to port 8888 on your host machine. Open a browser and go to `http://localhost:8888` to access the application running inside the Docker container and run the tutorials' notebooks.
 
-Now you can run the tutorials' notebook. More details about their content can be found [here](https://github.com/DeepRank/deeprank2/blob/main/tutorials/TUTORIAL.md). Note that in the docker container only the raw PDB files are downloaded, needed as a starting point for the tutorials. You can obtain the processed HDF5 files by running the `data_generation_xxx.ipynb` notebooks. Because Docker containers are limited in memory resources, we limit the number of data points processed in the tutorials'. Please install the package locally to fully leverage its capabilities.
[…]
