### 2023-12-11 #5983 @richardjgowers
@jnwei just checking in to see if the Dockerfile that has been added would address your review?
