### 2023-12-11 #5983 @jnwei
Yes, overall the Dockerfile in https://github.com/DeepRank/deeprank2/pull/528 looks great!

I'm quite swamped at the moment, but if you need someone to test the docker build in the new year, I could take closer look then.
