### 2023-12-12 599ba61
Update CITATION.cff
