+++ CITATION.cff
@@ -3,7 +3,7 @@
 
 cff-version: 1.2.0
 doi: 10.5281/zenodo.6772110
-title: Deeprank2
+title: DeepRank2
 message: >-
   If you use this software, please cite it using the
   metadata from this file.
