### 2023-12-18 #5983 @abhishektiwari
@gcroci2 Thanks for adding Dockerfile. I will be merging the PR anytime soon to mainline?
