### 2023-12-18 #5983 @abhishektiwari
@gcroci2

## Software Paper

Typos/minor corrections

— `publicly` not `publicily`
— `characterization` not `characterizartion`
— `classification` not `classfication`
— `Academic` not `Acamedic`
— `3D grids, graphs, and surfaces` not `3D grids, graphs and surfaces`

## Documentation
[Package Reference](https://deeprank2.readthedocs.io/en/latest/reference/deeprankcore.html) link on the index page/sidebar is missing. Please add them.
