### 2023-12-19 5583fea
use mambaforge instead of miniconda for make the env build faster
