### 2023-12-19 bbdf64c
handle cases in which losses are nan
