+++ deeprank2/trainer.py
@@ -577,6 +577,7 @@ class Trainer():
 
         train_losses = []
         valid_losses = []
+        checkpoint_model = None
 
         if earlystop_patience or earlystop_maxgap:
             early_stopping = EarlyStopping(patience=earlystop_patience, maxgap=earlystop_maxgap, min_epoch=min_epoch, trace_func=_log.info)
@@ -630,10 +631,13 @@ class Trainer():
                             _log.info(f'Best model saved at epoch # {self.epoch_saved_model}.')
 
             # Save the last model
-            if best_model is False:
+            if best_model is False or checkpoint_model is None:
                 checkpoint_model = self._save_model()
                 self.epoch_saved_model = epoch
                 _log.info(f'Last model saved at epoch # {self.epoch_saved_model}.')
+                if checkpoint_model is None:
+                    _log.warning("A model has been saved but the training and validation losses were NaN;" +
+                              "make sure that you are using enough data points during the trainig.")
 
         # Now that the training loop is over, save the model
         if filename:
