### 2023-12-19 992bc48
improve logic for warning about nan losses
