### 2023-12-20 50bbe82
user warnings for nan loss cases instead of logging
