### 2023-12-20 b6b6be7
add integration test for nan losses cases
