+++ tests/test_integration.py
@@ -1,4 +1,3 @@
-import glob
 import os
 import warnings
 from shutil import rmtree
