+++ CITATION.cff
@@ -56,4 +56,4 @@ keywords:
 license: Apache-2.0
 commit: 4e8823758ba03f824b4281f5689cb6a335ab2f6c
 version: "2.1.2"
-date-released: '2023-12-23'
+date-released: '2023-12-21'
