+++ CITATION.cff
@@ -2,7 +2,7 @@
 # Visit https://bit.ly/cffinit to generate yours today!
 
 cff-version: 1.2.0
-doi: 10.5281/zenodo.8305000
+doi: 10.5281/zenodo.10417724
 title: DeepRank2
 message: >-
   If you use this software, please cite it using the
