### 2023-12-21 1b8f1f0
update integration tests with the new query api
