### 2023-12-21 484d580
update deeprank2 installation version
