+++ env/requirements.txt
@@ -3,4 +3,4 @@ torch_scatter==2.1.2
 torch_sparse==0.6.18
 torch_cluster==1.6.3
 torch_spline_conv==1.2.2
-deeprank2==2.1.1
+deeprank2==2.1.2
