### 2023-12-21 b0f87a2
remove package ref from index.rst
