### 2023-12-21 #5983 @gcroci2
Hi @abhishektiwari, thanks for your suggestions and review!

I corrected the typos, fixed the link on the docs, and improved the docs for adding a custom feature, I hope that now it will be easier to follow.
