+++ deeprank2/dataset.py
@@ -30,7 +30,6 @@ class DeeprankDataset(Dataset):
     def __init__(self, # pylint: disable=too-many-arguments
                  hdf5_path: str | list[str],
                  subset: list[str] | None,
-                 train: bool,
                  train_data: str | GridDataset | GraphDataset | None,
                  target: str | None,
                  target_transform: bool | None,
@@ -58,7 +57,6 @@ class DeeprankDataset(Dataset):
             raise TypeError(f"hdf5_path: unexpected type: {type(hdf5_path)}")
 
         self.subset = subset
-        self.train = train
         self.train_data = train_data
         self.target = target
         self.target_transform = target_transform
@@ -457,7 +455,6 @@ class GridDataset(DeeprankDataset):
         self,
         hdf5_path: str | list,
         subset: list[str] | None = None,
-        train: bool = True,
         train_data: str | GridDataset | None = None,
         features: list[str] | str | Literal["all"] | None = "all",
         target: str | None = None,
@@ -474,17 +471,14 @@ class GridDataset(DeeprankDataset):
         Args:
             hdf5_path (str | list): Path to .HDF5 file(s). For multiple .HDF5 files, insert the paths in a list. Defaults to None.
             subset (list[str] | None, optional): list of keys from .HDF5 file to include. Defaults to None (meaning include all).
-            train (bool, optional): Boolean flag to determine if the instance represents the training set.
-                It needs to be set to False only if the instance represents the validation or testing set.
-                Defaults to True.
             train_data (str | class:`GridDataset` | None, optional): data to inherit information from the training dataset or the pre-trained model.
-                It is considered only if `train` is False. In the latter case, `train_data` needs to be a dataset of the same class or
+                If None, the current dataset is considered as the training set. Otherwise, `train_data` needs to be a dataset of the same class or
                 the path of a DeepRank2 pre-trained model. If set, the parameters `features`, `target`, `traget_transform`, `task`, and `classes`
                 will be inherited from `train_data`.
                 Defaults to None.
             features (list[str] | str | Literal["all"] | None, optional): Consider all pre-computed features ("all") or some defined node features
                 (provide a list, example: ["res_type", "polarity", "bsa"]). The complete list can be found in `deeprank2.domain.gridstorage`.
-                Value will be ignored and inherited from `train_data` if `train` is set as False and `train_data` is assigned.
+                Value will be ignored and inherited from `train_data` if `train_data` is assigned.
                 Defaults to "all".
             target (str | None, optional): Default options are irmsd, lrmsd, fnat, binary, capri_class, and dockq. It can also be
                 a custom-defined target given to the Query class as input (see: `deeprank2.query`); in this case,
@@ -492,11 +486,11 @@ class GridDataset(DeeprankDataset):
                 Only numerical target variables are supported, not categorical.
                 If the latter is your case, please convert the categorical classes into
                 numerical class indices before defining the :class:`GraphDataset` instance.
-                Value will be ignored and inherited from `train_data` if `train` is set as False and `train_data` is assigned.
+                Value will be ignored and inherited from `train_data` if `train_data` is assigned.
                 Defaults to None.
             target_transform (bool, optional): Apply a log and then a sigmoid transformation to the target (for regression only).
                 This puts the target value between 0 and 1, and can result in a more uniform target distribution and speed up the optimization.
-                Value will be ignored and inherited from `train_data` if `train` is set as False and `train_data` is assigned.
+                Value will be ignored and inherited from `train_data` if `train_data` is assigned.
                 Defaults to False.
             target_filter (dict[str, str] | None, optional): Dictionary of type [target: cond] to filter the molecules.
[…]
