### 2024-01-03 d1da845
remove train refs from trainer.py
