### 2024-01-03 0bb3023
update tests with the new train_data logic
