+++ deeprank2/dataset.py
@@ -30,7 +30,7 @@ class DeeprankDataset(Dataset):
     def __init__(self, # pylint: disable=too-many-arguments
                  hdf5_path: str | list[str],
                  subset: list[str] | None,
-                 train_data: str | GridDataset | GraphDataset | None,
+                 train_source: str | GridDataset | GraphDataset | None,
                  target: str | None,
                  target_transform: bool | None,
                  target_filter: dict[str, str] | None,
@@ -57,7 +57,7 @@ class DeeprankDataset(Dataset):
             raise TypeError(f"hdf5_path: unexpected type: {type(hdf5_path)}")
 
         self.subset = subset
-        self.train_data = train_data
+        self.train_source = train_source
         self.target = target
         self.target_transform = target_transform
         self.target_filter = target_filter
@@ -85,12 +85,12 @@ class DeeprankDataset(Dataset):
     def _check_and_inherit_train(self, data_type: Union[GridDataset, GraphDataset], inherited_params):
         """Check if the pre-trained model or training set provided are valid for validation and/or testing, and inherit the parameters.
         """
-        if isinstance(self.train_data, str): # pylint: disable=too-many-nested-blocks
+        if isinstance(self.train_source, str): # pylint: disable=too-many-nested-blocks
             try:
                 if torch.cuda.is_available():
-                    data = torch.load(self.train_data)
+                    data = torch.load(self.train_source)
                 else:
-                    data = torch.load(self.train_data, map_location=torch.device('cpu'))
+                    data = torch.load(self.train_source, map_location=torch.device('cpu'))
                 if data["data_type"] is not data_type:
                     raise TypeError (f"""The pre-trained model has been trained with data of type {data["data_type"]}, but you are trying
                                             to define a {data_type}-class validation/testing dataset. Please provide a valid DeepRank2
@@ -105,15 +105,15 @@ class DeeprankDataset(Dataset):
                                 continue
                             key['transform'] = eval(key['transform']) # pylint: disable=eval-used
             except pickle.UnpicklingError as e:
-                raise ValueError("""The path provided to `train_data` is not a valid DeepRank2 pre-trained model.
+                raise ValueError("""The path provided to `train_source` is not a valid DeepRank2 pre-trained model.
                                         Please provide a valid path to a DeepRank2 pre-trained model.""") from e
-        elif isinstance(self.train_data, data_type):
-            data = self.train_data
+        elif isinstance(self.train_source, data_type):
+            data = self.train_source
             if data_type is GraphDataset:
-                self.train_means = self.train_data.means
-                self.train_devs = self.train_data.devs
+                self.train_means = self.train_source.means
+                self.train_devs = self.train_source.devs
         else:
-            raise TypeError(f"""The train data provided is type: {type(self.train_data)}
+            raise TypeError(f"""The train data provided is type: {type(self.train_source)}
                             Please provide a valid training {data_type} or the path to a valid DeepRank2 pre-trained model.""")
 
         #match parameters with the ones in the training set
@@ -455,7 +455,7 @@ class GridDataset(DeeprankDataset):
         self,
         hdf5_path: str | list,
         subset: list[str] | None = None,
-        train_data: str | GridDataset | None = None,
+        train_source: str | GridDataset | None = None,
         features: list[str] | str | Literal["all"] | None = "all",
         target: str | None = None,
         target_transform: bool = False,
@@ -471,14 +471,14 @@ class GridDataset(DeeprankDataset):
         Args:
             hdf5_path (str | list): Path to .HDF5 file(s). For multiple .HDF5 files, insert the paths in a list. Defaults to None.
             subset (list[str] | None, optional): list of keys from .HDF5 file to include. Defaults to None (meaning include all).
[…]
