+++ tests/test_integration.py
@@ -78,12 +78,12 @@ def test_cnn(): # pylint: disable=too-many-locals
 
         dataset_val = GridDataset(
             hdf5_path = hdf5_paths,
-            train_data = dataset_train,
+            train_source = dataset_train,
         )
 
         dataset_test = GridDataset(
             hdf5_path = hdf5_paths,
-            train_data = dataset_train,
+            train_source = dataset_train,
         )
 
         output_exporters = [HDF5OutputExporter(output_directory)]
@@ -159,13 +159,13 @@ def test_gnn(): # pylint: disable=too-many-locals
 
         dataset_val = GraphDataset(
             hdf5_path = hdf5_paths,
-            train_data = dataset_train,
+            train_source = dataset_train,
             clustering_method = "mcl"
         )
 
         dataset_test = GraphDataset(
             hdf5_path = hdf5_paths,
-            train_data = dataset_train,
+            train_source = dataset_train,
             clustering_method = "mcl"
         )
 

[…]
