### 2024-01-09 #5983 @jnwei
Hi @gcroci2 ,

Thank you for your work on the Docker setup! This is great, I was able to follow the instructions to set up both the docker container and run the tutorial notebooks. I have completed the reproducibility sections of the review checklist accordingly.

One note, when I created a sample dataset with either data creation notebook (e.g. `data_generation_ppi.ipynb` or `data_generation_srv.ipynb` ), then attempted training with the `training.ipynb`, I observed that losses for the training of the GNN model were all nans.  The CNN model ran correctly, with the training/validation loss decreasing over number of training steps. I am not sure if this is the intended behavior or not.
