### 2024-01-15 4fb9f8e
add conda installation for dssp in the yml file
