### 2024-01-15 f9f93bc
remove the not-conda dssp installation
