+++ .github/actions/install-python-and-package/action.yml
@@ -44,8 +44,6 @@ runs:
       if: runner.os == 'Linux'
       run: |
         # Install dependencies not handled by setuptools 
-        ## DSSP
-        sudo apt-get install -y dssp
         ## MSMS
         conda install -c bioconda msms
         ## PyTorch, PyG, PyG adds
@@ -60,27 +58,6 @@ runs:
       if: runner.os == 'macOS'
       run: |
         # Install dependencies not handled by setuptools 
-        ## DSSP
-        git clone https://github.com/PDB-REDO/libcifpp.git --recurse-submodules
-        cd libcifpp
-        cmake -S . -B build -DCMAKE_INSTALL_PREFIX=$HOME/.local -DCMAKE_BUILD_TYPE=Release
-        cmake --build build
-        cmake --install build
-        #######
-        git clone https://github.com/mhekkel/libmcfp.git
-        cd libmcfp
-        mkdir build
-        cd build
-        cmake ..
-        cmake --build .
-        cmake --install .
-        #######
-        git clone https://github.com/PDB-REDO/dssp.git
-        cd dssp
-        mkdir build
-        cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
-        cmake --build build
-        cmake --install build
         ## MSMS
         cd /tmp/
         wget http://mgltools.scripps.edu/downloads/tars/releases/MSMSRELEASE/REL2.6.1/msms_i86Linux2_2.6.1.tar.gz
