### 2024-01-15 e106c88
add dssp conda installation to the action.yml file
