+++ .github/actions/install-python-and-package/action.yml
@@ -44,6 +44,8 @@ runs:
       if: runner.os == 'Linux'
       run: |
         # Install dependencies not handled by setuptools 
+        ## DSSP 
+        conda install sbl::dssp
         ## MSMS
         conda install -c bioconda msms
         ## PyTorch, PyG, PyG adds
@@ -58,6 +60,8 @@ runs:
       if: runner.os == 'macOS'
       run: |
         # Install dependencies not handled by setuptools 
+        ## DSSP 
+        conda install sbl::dssp
         ## MSMS
         cd /tmp/
         wget http://mgltools.scripps.edu/downloads/tars/releases/MSMSRELEASE/REL2.6.1/msms_i86Linux2_2.6.1.tar.gz
