### 2024-01-15 fc57c94
try to fix dssp conda installation
