### 2024-01-15 4f0a478
add libgcc-ng to the conda installation
