### 2024-01-16 b5767f6
ci: set up ruff linter
