### 2024-01-16 01ccd90
docs: reference new linter in documentation
