### 2024-01-16 bf66cac
style: implement new linter/formatter throughout code base
