+++ deeprank2/features/secondary_structure.py
@@ -12,15 +12,15 @@ from deeprank2.utils.graph import Graph
 
 
 class DSSPError(Exception):
-    "Raised if DSSP fails to produce an output"
+    """Raised if DSSP fails to produce an output."""
 
 
 class SecondarySctructure(Enum):
-    "a value to express a secondary a residue's secondary structure type"
+    """Value to express a secondary a residue's secondary structure type."""
 
-    HELIX = 0 # 'GHI'
-    STRAND = 1 # 'BE'
-    COIL = 2 # ' -STP'
+    HELIX = 0  # 'GHI'
+    STRAND = 1  # 'BE'
+    COIL = 2  # ' -STP'
 
     @property
     def onehot(self):
@@ -38,45 +38,45 @@ def _get_records(lines: list[str]):
 
 def _check_pdb(pdb_path: str):
     fix_pdb = False
-    with open(pdb_path, encoding='utf-8') as f:
+    with open(pdb_path, encoding="utf-8") as f:
         lines = f.readlines()
 
     # check for HEADER
     firstline = lines[0]
-    if not firstline.startswith('HEADER'):
+    if not firstline.startswith("HEADER"):
         fix_pdb = True
-        if firstline.startswith('EXPDTA'):
-            lines = [f'HEADER {firstline}'] + lines[1:]
+        if firstline.startswith("EXPDTA"):
+            lines = [f"HEADER {firstline}"] + lines[1:]
         else:
-            lines = ['HEADER \n'] + lines
+            lines = ["HEADER \n", *lines]
 
     # check for CRYST1 record
     existing_records = _get_records(lines)
-    if 'CRYST1' not in existing_records:
+    if "CRYST1" not in existing_records:
         fix_pdb = True
-        dummy_CRYST1 = 'CRYST1   00.000   00.000   00.000  00.00  00.00  00.00 X 00 00 0    00\n'
+        dummy_CRYST1 = "CRYST1   00.000   00.000   00.000  00.00  00.00  00.00 X 00 00 0    00\n"
         lines = [lines[0]] + [dummy_CRYST1] + lines[1:]
 
     # check for unnumbered REMARK lines
     for i, line in enumerate(lines):
-        if line.startswith('REMARK'):
+        if line.startswith("REMARK"):
             try:
                 int(line.split()[1])
             except ValueError:
                 fix_pdb = True
-                lines[i] = f'REMARK 999 {line[7:]}'
+                lines[i] = f"REMARK 999 {line[7:]}"
 
     if fix_pdb:
-        with open(pdb_path, 'w', encoding='utf-8') as f:
+        with open(pdb_path, "w", encoding="utf-8") as f:
             f.writelines(lines)
 
 
 def _classify_secstructure(subtype: str):
-    if subtype in 'GHI':
+    if subtype in "GHI":
         return SecondarySctructure.HELIX
-    if subtype in 'BE':
+    if subtype in "BE":
         return SecondarySctructure.STRAND
-    if subtype in ' -STP':
+    if subtype in " -STP":
         return SecondarySctructure.COIL
     return None
 
@@ -90,21 +90,21 @@ def _get_secstructure(pdb_path: str) -> dict:
     Returns:
         dict: A dictionary containing secondary structure information for each chain and residue.
     """
-
     # Execute DSSP and read the output
     _check_pdb(pdb_path)
     p = PDBParser(QUIET=True)
     model = p.get_structure(Path(pdb_path).stem, pdb_path)[0]
 
-    # pylint: disable=raise-missing-from
     try:
-        dssp = DSSP(model, pdb_path, dssp='mkdssp')
[…]
