+++ deeprank2/utils/parsing/top.py
@@ -4,6 +4,7 @@ from typing import Any
 
 logging.getLogger(__name__)
 
+
 class TopRowObject:
     def __init__(
         self,
@@ -18,11 +19,10 @@ class TopRowObject:
     def __getitem__(self, key):
         return self.kwargs[key]
 
+
 class TopParser:
     _VAR_PATTERN = re.compile(r"([^\s]+)\s*=\s*([^\s\(\)]+|\(.*\))")
-    _LINE_PATTERN = re.compile(
-        r"^([A-Z0-9]{3})\s+atom\s+([A-Z0-9]{1,4})\s+(.+)\s+end\s*(\s+\!\s+[ _A-Za-z0-9]+)?$"
-    )
+    _LINE_PATTERN = re.compile(r"^([A-Z0-9]{3})\s+atom\s+([A-Z0-9]{1,4})\s+(.+)\s+end\s*(\s+\!\s+[ _A-Za-z0-9]+)?$")
     _NUMBER_PATTERN = re.compile(r"\-?[0-9]+(\.[0-9]+)?")
 
     @staticmethod
@@ -39,9 +39,7 @@ class TopParser:
 
             kwargs = {}
             for w in TopParser._VAR_PATTERN.finditer(m.group(3)):
-                kwargs[w.group(1).lower().strip()] = TopParser._parse_value(
-                    w.group(2).strip()
-                )
+                kwargs[w.group(1).lower().strip()] = TopParser._parse_value(w.group(2).strip())
 
             result.append(TopRowObject(residue_name, atom_name, kwargs))
 
