### 2024-01-16 054c7d3
test: suppress some warnings
