+++ deeprank2/trainer.py
@@ -813,10 +813,6 @@ class Trainer:
             target = torch.tensor([self.classes_to_index[x] if isinstance(x, str) else self.classes_to_index[int(x)] for x in target])
             if isinstance(self.lossfunction, nn.BCELoss | nn.BCEWithLogitsLoss):
                 # # pred must be in (0,1) range and target must be float with same shape as pred
-                # pred = F.softmax(pred)
-                # target = torch.tensor(
-                #     [[0,1] if x == [1] else [1,0] for x in target]
-                # ).float()
                 raise ValueError(
                     "BCELoss and BCEWithLogitsLoss are currently not supported.\n\t"
                     "For further details see: https://github.com/DeepRank/deeprank2/issues/318"
