+++ tests/features/test_secondary_structure.py
@@ -33,12 +33,12 @@ def test_secondary_structure_residue():
         (267, "A", " ", SecondarySctructure.COIL),
         (46, "A", "S", SecondarySctructure.COIL),
         (104, "A", "T", SecondarySctructure.COIL),
-        # (None, '', 'P', SecondarySctructure.COIL), # not found in test file
+        # (None, '', 'P', SecondarySctructure.COIL),  # not found in test file  # noqa: ERA001 (commented-out code)
         (194, "A", "B", SecondarySctructure.STRAND),
         (385, "B", "E", SecondarySctructure.STRAND),
         (235, "A", "G", SecondarySctructure.HELIX),
         (263, "A", "H", SecondarySctructure.HELIX),
-        # (0, '', 'I', SecondarySctructure.HELIX), # not found in test file
+        # (0, '', 'I', SecondarySctructure.HELIX),  # not found in test file  # noqa: ERA001 (commented-out code)
     ]
 
     for res in residues:
