+++ tests/utils/test_earlystopping.py
@@ -25,6 +25,7 @@ def base_earlystopper(patience=10, delta=0, maxgap=None):
 def test_patience():
     patience = 3
     final_ep = base_earlystopper(patience=patience)
+    # should terminate at epoch 7
     assert final_ep == 7
 
 
@@ -32,15 +33,12 @@ def test_patience_with_delta():
     patience = 3
     delta = 1
     final_ep = base_earlystopper(patience=patience, delta=delta)
+    # should terminate at epoch 5
     assert final_ep == 5
 
 
 def test_maxgap():
     maxgap = 1
     final_ep = base_earlystopper(maxgap=maxgap)
+    # should terminate at epoch 9
     assert final_ep == 9
-
-
-# test_patience() # should terminate at epoch 7
-# test_patience_with_delta() # should terminate at epoch 5
-# test_maxgap() # should terminate at epoch 9
