+++ .github/workflows/markdown-link-check.yml
@@ -3,7 +3,7 @@ name: markdown-link-check
 on:
   push:
     branches:
-    - main
+      - main
     paths:
       # filetypes
       - "**.md"
@@ -25,7 +25,7 @@ jobs:
     name: Check markdown links
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
-    - uses: gaurav-nelson/github-action-markdown-link-check@v1
-      with:
-        config-file: '.github/workflows/markdown-link-check.yml'
+      - uses: actions/checkout@v3
+      - uses: gaurav-nelson/github-action-markdown-link-check@v1
+        with:
+          config-file: ".github/workflows/markdown-link-check.yml"
