+++ .github/workflows/release.yml
@@ -14,8 +14,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-latest']
-        python-version: ['3.10']
+        os: ["ubuntu-latest"]
+        python-version: ["3.10"]
 
     steps:
       - uses: actions/checkout@v3

[…]
