### 2024-01-18 93e7a5a
try to print out warning for failing tests
