+++ .github/workflows/build.yml
@@ -46,6 +46,6 @@ jobs:
           python-version: ${{ matrix.python-version }}
           extras-require: test, publishing
       - name: Run unit tests
-        run: pytest -v
+        run: pytest -v -s
       - name: Verify that we can build the package
         run: python3 -m build
