+++ env/environment.yml
@@ -5,6 +5,7 @@ channels:
   - bioconda
   - defaults
   - sbl
+  - conda-forge
 dependencies:
   - pip==23.3.*
   - python==3.10.*
