### 2024-01-18 91775b6
try to fix conda activate
