### 2024-01-18 0faf301
try to fix conda env installation
