+++ .github/actions/install-python-and-package/action.yml
@@ -25,52 +25,20 @@ runs:
           access_token: ${{ github.token }}
     - uses: actions/checkout@v3
     - name: Setup conda
-      uses: s-weigand/setup-conda@v1
+      uses: conda-incubator/setup-miniconda@v2
       with:
-        update-conda: true
-        python-version: ${{ inputs.python-version }}
-        # conda-channels: anaconda
-    - run: conda --version
+        activate-environment: deeprank2
+        environment-file: ./env/environment.yml
+        auto-activate-base: false
+    - run: |
+        conda info
+        conda list
       shell: bash {0}
     - name: Python info
       shell: bash -e {0}
       run: |
         which python3
         python3 --version
-    - name: Install dependencies on Linux
-      shell: bash {0}
-      env:
-        CMAKE_INSTALL_PREFIX: .local
-      if: runner.os == 'Linux'
-      run: |
-        conda env create -f ./env/environment.yml
-        conda activate deeprank2 || true
-        conda init bash
-    - name: Install dependencies on MacOS
-      shell: bash {0}
-      env:
-        CMAKE_INSTALL_PREFIX: .local
-      if: runner.os == 'macOS'
-      run: |
-        # Install dependencies not handled by setuptools 
-        conda install -c conda-forge libgcc-ng
-        ## DSSP 
-        conda install -c sbl dssp
-        ## MSMS
-        cd /tmp/
-        wget http://mgltools.scripps.edu/downloads/tars/releases/MSMSRELEASE/REL2.6.1/msms_i86Linux2_2.6.1.tar.gz
-        sudo mkdir /usr/local/lib/msms
-        cd /usr/local/lib/msms
-        sudo tar zxvf /tmp/msms_i86Linux2_2.6.1.tar.gz
-        sudo ln -s /usr/local/lib/msms/msms.i86Linux2.2.6.1 /usr/local/bin/msms
-        sudo ln -s /usr/local/lib/msms/pdb_to_xyzr* /usr/local/bin
-        ## PyTorch, PyG, PyG adds
-        ### Installing for CPU only on the CI
-        conda install pytorch torchvision torchaudio cpuonly -c pytorch
-        pip install torch_geometric
-        pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-$(python3 -c "import torch; print(torch.__version__)")+cpu.html
-        # PyTables via conda only for MacOS
-        conda install pytables
     - name: Install the package
       shell: bash {0}
       run: pip install .'[${{ inputs.extras-require }}]'
