### 2024-01-18 941823b
update installation on CI
