+++ .github/actions/install-python-and-package/action.yml
@@ -25,20 +25,26 @@ runs:
           access_token: ${{ github.token }}
     - uses: actions/checkout@v3
     - name: Setup conda
-      uses: conda-incubator/setup-miniconda@v2
+      uses: s-weigand/setup-conda@v1
       with:
-        activate-environment: deeprank2
-        environment-file: ./env/environment.yml
-        auto-activate-base: false
-    - run: |
-        conda info
-        conda list
+        update-conda: true
+        python-version: ${{ inputs.python-version }}
+        conda-channels: pytorch, pyg, bioconda, defaults, sbl, conda-forge
+    - run: conda --version
       shell: bash {0}
     - name: Python info
       shell: bash -e {0}
       run: |
         which python3
         python3 --version
+    - name: Install dependencies on Linux
+      shell: bash {0}
+      env:
+        CMAKE_INSTALL_PREFIX: .local
+      if: runner.os == 'Linux'
+      run: |
+        # Install dependencies
+        conda env update --file ./env/environment.yml
     - name: Install the package
       shell: bash {0}
       run: pip install .'[${{ inputs.extras-require }}]'
