+++ .github/actions/install-python-and-package/action.yml
@@ -18,6 +18,10 @@ runs:
 
   using: "composite"
 
+  defaults:
+    run:
+      shell: bash -el {0}
+      
   steps:
     - name: Cancel Previous Runs and Set up Python
       uses: styfle/cancel-workflow-action@0.4.0
@@ -25,11 +29,12 @@ runs:
           access_token: ${{ github.token }}
     - uses: actions/checkout@v3
     - name: Setup conda
-      uses: s-weigand/setup-conda@v1
+      uses: conda-incubator/setup-miniconda@v3
       with:
-        update-conda: true
+        auto-update-conda: true
         python-version: ${{ inputs.python-version }}
-        conda-channels: pytorch, pyg, bioconda, defaults, sbl, conda-forge
+        activate-environment: deeprank2
+        environment-file: env/environment.yml
     - run: conda --version
       shell: bash {0}
     - name: Python info
