+++ env/environment.yml
@@ -8,7 +8,6 @@ channels:
   - conda-forge
 dependencies:
   - pip==23.3.*
-  - python==3.10.*
   - msms==2.6.1
   - pytorch==2.1.1
   - pytorch-mutex==1.0.*
