### 2024-01-19 d7f05c5
print info about conda env
