+++ .github/actions/install-python-and-package/action.yml
@@ -37,19 +37,15 @@ runs:
         environment-file: env/environment.yml
     - run: conda --version
       shell: bash {0}
+    - run: |
+        conda env list
+        conda list
+      shell: bash -el {0}
     - name: Python info
       shell: bash -e {0}
       run: |
         which python3
         python3 --version
-    - name: Install dependencies on Linux
-      shell: bash {0}
-      env:
-        CMAKE_INSTALL_PREFIX: .local
-      if: runner.os == 'Linux'
-      run: |
-        # Install dependencies
-        conda env update --file ./env/environment.yml
     - name: Install the package
       shell: bash {0}
       run: pip install .'[${{ inputs.extras-require }}]'
