### 2024-01-19 5f902ca
go back to original action but with dssp installed via conda
