+++ .github/actions/install-python-and-package/action.yml
@@ -29,23 +29,34 @@ runs:
           access_token: ${{ github.token }}
     - uses: actions/checkout@v3
     - name: Setup conda
-      uses: conda-incubator/setup-miniconda@v3
+      uses: s-weigand/setup-conda@v1
       with:
-        auto-update-conda: true
+        update-conda: true
         python-version: ${{ inputs.python-version }}
-        activate-environment: deeprank2
-        environment-file: env/environment.yml
+        conda-channels: pytorch, pyg, bioconda, defaults, sbl, conda-forge
     - run: conda --version
       shell: bash {0}
-    - run: |
-        conda env list
-        conda list
-      shell: bash -el {0}
     - name: Python info
       shell: bash -e {0}
       run: |
         which python3
         python3 --version
+    - name: Install dependencies on Linux
+      shell: bash {0}
+      env:
+        CMAKE_INSTALL_PREFIX: .local
+      if: runner.os == 'Linux'
+      run: |
+        # Install dependencies not handled by setuptools 
+        ## DSSP
+        conda install -c sbl dssp=4.2.2.1
+        ## MSMS
+        conda install -c bioconda msms=2.6.1
+        ## PyTorch, PyG, PyG adds
+        ### Installing for CPU only on the CI
+        conda install pytorch=2.1.1 torchvision=0.16.1 torchaudio=2.1.1 cpuonly=2.0.* -c pytorch
+        conda install pyg=2.4.0 -c pyg
+        pip install torch_scatter==2.1.2 torch_sparse==0.6.18 torch_cluster==1.6.3 torch_spline_conv==1.2.2 -f https://data.pyg.org/whl/torch-2.1.0+cpu.html
     - name: Install the package
       shell: bash {0}
       run: pip install .'[${{ inputs.extras-require }}]'
