### 2024-01-19 7ecbe7d
add pip dep again to the toml
