### 2024-01-19 5b8dd7a
add conda env list
