+++ pyproject.toml
@@ -34,24 +34,24 @@ classifiers = [
     "Programming Language :: Python :: 3.10"
 ]
 
-dependencies = [
-    "tables >= 3.8.0",
-    "numpy >= 1.21.5",
-    "scipy >= 1.11.2",
-    "h5py >= 3.6.0",
-    "networkx >= 2.6.3",
-    "matplotlib >= 3.5.1",
-    "pdb2sql >= 0.5.1",
-    "scikit-learn >= 1.0.2",
-    "chart-studio >= 1.1.0",
-    "biopython >= 1.81",
-    "python-louvain >= 0.16",
-    "markov-clustering >= 0.0.6.dev0",
-    "tqdm >= 4.63.0",
-    "freesasa >= 2.1.0",
-    "tensorboard >= 0.9.0",
-    "protobuf >= 3.20.1"
-]
+# dependencies = [
+#     "tables >= 3.8.0",
+#     "numpy >= 1.21.5",
+#     "scipy >= 1.11.2",
+#     "h5py >= 3.6.0",
+#     "networkx >= 2.6.3",
+#     "matplotlib >= 3.5.1",
+#     "pdb2sql >= 0.5.1",
+#     "scikit-learn >= 1.0.2",
+#     "chart-studio >= 1.1.0",
+#     "biopython >= 1.81",
+#     "python-louvain >= 0.16",
+#     "markov-clustering >= 0.0.6.dev0",
+#     "tqdm >= 4.63.0",
+#     "freesasa >= 2.1.0",
+#     "tensorboard >= 0.9.0",
+#     "protobuf >= 3.20.1"
+# ]
 
 [project.optional-dependencies]
 # development dependency groups
