### 2024-01-19 b777787
add h5py to yml
