+++ env/environment.yml
@@ -17,5 +17,6 @@ dependencies:
   - pyg==2.4.0
   - notebook==7.0.6
   - dssp==4.2.2.1
+  - h5py>=3.6.0
   - pip:
     - --requirement requirements.txt
