### 2024-01-19 141b981
add missing deps to the yml file
