+++ env/environment.yml
@@ -9,6 +9,8 @@ channels:
 dependencies:
   - pip==23.3.*
   - msms==2.6.1
+  - dssp==4.2.2.1
+  - biopython >= 1.81
   - pytorch==2.1.1
   - pytorch-mutex==1.0.*
   - torchvision==0.16.1
@@ -16,7 +18,19 @@ dependencies:
   - cpuonly==2.0.*
   - pyg==2.4.0
   - notebook==7.0.6
-  - dssp==4.2.2.1
   - h5py>=3.6.0
+  - numpy>=1.21.5
+  - scipy>=1.11.2
+  - tables>=3.8.0
+  - matplotlib>=3.5.1
+  - pdb2sql >= 0.5.1
+  - scikit-learn >= 1.0.2
+  - chart-studio >= 1.1.0
+  - python-louvain >= 0.16
+  - markov-clustering >= 0.0.6.dev0
+  - tqdm >= 4.63.0
+  - freesasa >= 2.1.0
+  - tensorboard >= 0.9.0
+  - protobuf >= 3.20.1
   - pip:
     - --requirement requirements.txt
