+++ env/environment.yml
@@ -23,11 +23,10 @@ dependencies:
   - scipy>=1.11.2
   - tables>=3.8.0
   - matplotlib>=3.5.1
-  - pdb2sql >= 0.5.1
   - scikit-learn >= 1.0.2
   - chart-studio >= 1.1.0
   - python-louvain >= 0.16
-  - markov-clustering >= 0.0.6.dev0
+  - markov-clustering >= 0.0.6
   - tqdm >= 4.63.0
   - freesasa >= 2.1.0
   - tensorboard >= 0.9.0
