### 2024-01-19 c551e47
fix markov-clustering installation
