+++ env/environment.yml
@@ -26,7 +26,7 @@ dependencies:
   - scikit-learn >= 1.0.2
   - chart-studio >= 1.1.0
   - python-louvain >= 0.16
-  - markov-clustering >= 0.0.6
+  - markov_clustering >= 0.0.6
   - tqdm >= 4.63.0
   - freesasa >= 2.1.0
   - tensorboard >= 0.9.0
