+++ env/environment.yml
@@ -1,4 +1,3 @@
-name: deeprank2
 channels:
   - pytorch
   - pyg
