### 2024-01-19 144d817
put dependencies back to the toml
