+++ env/environment.yml
@@ -8,6 +8,7 @@ channels:
   - conda-forge
 dependencies:
   - pip==23.3.*
+  - python==3.10.*
   - msms==2.6.1
   - dssp==4.2.2.1
   - pytorch==2.1.1
