### 2024-01-19 986b6dd
readd macos deps installation
