+++ .github/actions/install-python-and-package/action.yml
@@ -47,14 +47,38 @@ runs:
       run: |
         # Install deeprank2 conda dependencies
         ## DSSP
-        conda install -c sbl dssp=4.2.2.1
+        conda install -c sbl dssp>=4.2.2.1
         ## MSMS
-        conda install -c bioconda msms=2.6.1
+        conda install -c bioconda msms>=2.6.1
         ## PyTorch, PyG, PyG adds
         ### Installing for CPU only on the CI
         conda install pytorch=2.1.1 torchvision=0.16.1 torchaudio=2.1.1 cpuonly=2.0.* -c pytorch
         conda install pyg=2.4.0 -c pyg
         pip install torch_scatter==2.1.2 torch_sparse==0.6.18 torch_cluster==1.6.3 torch_spline_conv==1.2.2 -f https://data.pyg.org/whl/torch-2.1.0+cpu.html
+    - name: Install dependencies on MacOS
+      shell: bash {0}
+      env:
+        CMAKE_INSTALL_PREFIX: .local
+      if: runner.os == 'macOS'
+      run: |
+        # Install dependencies not handled by setuptools 
+        ## DSSP
+        conda install -c sbl dssp>=4.2.2.1
+        ## MSMS
+        cd /tmp/
+        wget http://mgltools.scripps.edu/downloads/tars/releases/MSMSRELEASE/REL2.6.1/msms_i86Linux2_2.6.1.tar.gz
+        sudo mkdir /usr/local/lib/msms
+        cd /usr/local/lib/msms
+        sudo tar zxvf /tmp/msms_i86Linux2_2.6.1.tar.gz
+        sudo ln -s /usr/local/lib/msms/msms.i86Linux2.2.6.1 /usr/local/bin/msms
+        sudo ln -s /usr/local/lib/msms/pdb_to_xyzr* /usr/local/bin
+        ## PyTorch, PyG, PyG adds
+        ### Installing for CPU only on the CI
+        conda install pytorch torchvision torchaudio cpuonly -c pytorch
+        pip install torch_geometric
+        pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-$(python3 -c "import torch; print(torch.__version__)")+cpu.html
+        # PyTables via conda only for MacOS
+        conda install pytables
     - name: Install the package
       shell: bash {0}
       run: pip install .'[${{ inputs.extras-require }}]'
