### 2024-01-19 c41f132
remove dssp from yml - the dockerfile does not need to be edited at this stage
