+++ env/environment.yml
@@ -4,13 +4,11 @@ channels:
   - pyg
   - bioconda
   - defaults
-  - sbl
   - conda-forge
 dependencies:
   - pip==23.3.*
   - python==3.10.*
   - msms==2.6.1
-  - dssp==4.2.2.1
   - pytorch==2.1.1
   - pytorch-mutex==1.0.*
   - torchvision==0.16.1
