### 2024-01-19 d8336aa
update docker file with dssp via conda
