### 2024-01-19 de7ad35
update docs for new dssp installation
