+++ paper/paper.bib
@@ -31,7 +31,7 @@
 	volume = {21},
 	issn = {0969-2126},
 	url = {https://www.sciencedirect.com/science/article/pii/S0969212613002992},
-	doi = {https://doi.org/10.1016/j.str.2013.08.007},
+	doi = {10.1016/j.str.2013.08.007},
 	pages = {1531--1540},
 	number = {9},
 	journaltitle = {Structure},
@@ -47,7 +47,7 @@ number = {4},
 pages = {401-404},
 year = {2000},
 issn = {0959-440X},
-doi = {https://doi.org/10.1016/S0959-440X(00)00105-6},
+doi = {10.1016/S0959-440X(00)00105-6},
 url = {https://www.sciencedirect.com/science/article/pii/S0959440X00001056},
 author = {Paul J Gane and Philip M Dean},
 keywords = {Structure-based drug design, Docking, Ligand binding, Flexible docking, Scoring, High-throughput screening}
@@ -125,6 +125,7 @@ URL = {https://www.biorxiv.org/content/10.1101/2021.10.04.463034v2}
   number={7},
   pages={2113--2118},
   year={2020},
+  doi={10.1093/bioinformatics/btz870},
   publisher={Oxford University Press}
 }
 
@@ -164,7 +165,7 @@ volume = {89},
 number = {5},
 pages = {493-501},
 keywords = {equivariant neural network, physics-aware machine learning, protein docking, representation learning},
-doi = {https://doi.org/10.1002/prot.26033},
+doi = {10.1002/prot.26033},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/prot.26033},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/prot.26033},
 year = {2021}
@@ -199,6 +200,7 @@ note ={PMCID: PMC7728386}
   journal={Frontiers in Molecular Biosciences},
   volume={10},
   year={2023},
+  doi={10.3389/fmolb.2023.1204157},
   publisher={Frontiers Media SA}
 }
 
