### 2024-01-23 #5983 @gcroci2
> @gcroci2 there's a couple of DOI fixes required (listed above) but then it looks like we're nearly finished here.

Now they should be fixed, thanks!
