### 2024-01-23 34c480e
minor improvements for authors and affiliations
