+++ CITATION.cff
@@ -3,7 +3,7 @@
 
 cff-version: 1.2.0
 doi: 10.5281/zenodo.10417724
-title: DeepRank2
+title: 'DeepRank2: Mining 3D Protein Structures with Geometric Deep Learning'
 message: >-
   If you use this software, please cite it using the
   metadata from this file.
@@ -12,46 +12,59 @@ authors:
   - given-names: Giulia
     family-names: Crocioni
     email: g.crocioni@esciencecenter.nl
-    affiliation: Netherlands eScience Center
+    affiliation: Netherlands eScience Center, Amsterdam, NL
     orcid: 'https://orcid.org/0000-0002-0823-0121'
-  - given-names: Dani
+  - given-names: Dani L.
     family-names: Bodor
     email: d.bodor@esciencecenter.nl
-    affiliation: Netherlands eScience Center
+    affiliation: Netherlands eScience Center, Amsterdam, NL
     orcid: 'https://orcid.org/0000-0003-2109-2349'
   - given-names: Coos
     family-names: Baakman
-    affiliation: Radboud University Medical Center
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
     orcid: 'https://orcid.org/0000-0003-4317-1566'
-  - given-names: Daniel
+  - given-names: Farzaneh Meimandi
+    family-names: Parizi
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
+    orcid: 'https://orcid.org/0000-0003-4230-7492'
+  - given-names: Daniel-T.
     family-names: Rademaker
-    affiliation: Radboud University Medical Center
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
     orcid: 'https://orcid.org/0000-0003-1959-1317'
   - given-names: Gayatri
     family-names: Ramakrishnan
-    affiliation: Radboud University Medical Center
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
     orcid: 'https://orcid.org/0000-0001-8203-2783'
-  - given-names: Sven
+  - given-names: Sven A.
     family-names: van der Burg
-    affiliation: Netherlands eScience Center
+    affiliation: Netherlands eScience Center, Amsterdam, NL
     orcid: 'https://orcid.org/0000-0003-1250-6968'
-  - given-names: Farzaneh Meimandi
-    family-names: Parizi
-    affiliation: Radboud University Medical Center
-    orcid: 'https://orcid.org/0000-0003-4230-7492'
+  - given-names: Dario F.
+    family-names: Marzella
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
+    orcid: 'https://orcid.org/0000-0002-0043-3055'
+  - given-names: João M.C.
+    family-names: Teixeira
+    affiliation: Independent Researcher
+    orcid: 'https://orcid.org/0000-0002-9113-0622'
   - given-names: Li C.
     family-names: Xue
-    affiliation: Radboud University Medical Center
+    affiliation: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
     orcid: 'https://orcid.org/0000-0002-2613-538X'
 abstract: >-
   DeepRank2 is an open-source deep learning
   framework for data mining of protein-protein interfaces
   or single-residue variants.
 keywords:
-  - Graph neural networks
-  - Convolutional neural networks
[…]
