### 2024-01-24 28b9735
fix typos and edit subtitles
