### 2024-01-24 7a4ce85
add toc to docs/installation.md
