### 2024-01-24 08a4013
remove pyhon 3.11 from the CI
