### 2024-01-24 0b63729
add extra input for testing release in the action.yml
