+++ .github/workflows/coveralls.yml
@@ -45,6 +45,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           extras-require: test
+          test-release: false
       - name: Run unit tests with coverage
         run: pytest --cov --cov-append --cov-report xml --cov-report term --cov-report html
       - name: Coveralls
