+++ .github/actions/install-python-and-package/action.yml
@@ -13,6 +13,11 @@ inputs:
     description: "The extras dependencies packages to be installed, for instance 'dev' or 'dev,publishing,notebooks'."
     default: "test"
 
+  test-release:
+    required: false
+    description: "Boolean to specify if testing the latest released version of the package (true) or the editable one (false)."
+    default: false
+
 runs:
   using: "composite"
 
@@ -94,6 +99,11 @@ runs:
         pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-$(python3 -c "import torch; print(torch.__version__)")+cpu.html
         # PyTables via conda only for MacOS
         conda install pytables
-    - name: Install the package
+    - name: Install the editable version of the package
       shell: bash {0}
+      if: ${{ inputs.test-release }} == false
       run: pip install .'[${{ inputs.extras-require }}]'
+    - name: Install the latest released version of the package
+      shell: bash {0}
+      if: ${{ inputs.test-release }} == true
+      run: pip install deeprank2
