### 2024-01-24 ac12feb
Update docs/installation.md
Co-authored-by: Dani Bodor <d.bodor@esciencecenter.nl>
