### 2024-01-24 f1519d2
Update docs/installation.md
Co-authored-by: Dani Bodor <d.bodor@esciencecenter.nl>
