### 2024-01-24 386e6ff
add yml file for testing latest release
