+++ .github/workflows/build-latest-release.yml
@@ -1,4 +1,4 @@
-name: build
+name: build (latest release)
 
 on:
   push:
