### 2024-01-24 5f24eef
remove deeprank2 dir from env before installing latest release
