### 2024-01-24 f9ecec7
try to fix test-release input format
