+++ .github/workflows/build-latest-release.yml
@@ -45,7 +45,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           extras-require: test, publishing
-          test-release: true
+          test-release: "true"
       - name: Run unit tests
         run: pytest -v
       - name: Verify that we can build the package
