### 2024-01-24 bba34e3
try to fix if cond
