### 2024-01-24 3dc2d99
use string for if cond on test package input
