### 2024-01-24 0a4ae17
add pytest in latest version case
