### 2024-01-24 967e436
try to fix pytest installation
