### 2024-01-25 974cb60
make build latest release action run only when a release action is completed
