### 2024-01-25 469139f
rename test-release input to pkg-installation-type
