### 2024-01-25 bdf94d8
improve description for pkg-installation-type
