+++ .github/actions/install-python-and-package/action.yml
@@ -15,7 +15,7 @@ inputs:
 
   pkg-installation-type:
     required: false
-    description: "To specify if testing the latest released version of the package (true) or the editable one (false)."
+    description: "The package installation type to install, the latest released version (latest) or the editable one (editable)."
     default: "editable"
 
 runs:
