+++ .github/workflows/build-repo.yml
@@ -1,4 +1,4 @@
-name: build (editable package)
+name: build (repository package)
 
 on:
   push:
@@ -45,7 +45,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           extras-require: test, publishing
-          pkg-installation-type: "editable"
+          pkg-installation-type: "repository"
       - name: Run unit tests
         run: pytest -v
       - name: Verify that we can build the package
