+++ .github/workflows/release.yml
@@ -23,7 +23,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           extras-require: publishing
-          pkg-installation-type: "editable"
+          pkg-installation-type: "repository"
       - name: Build wheel and source distribution
         run: python -m build
       - uses: actions/upload-artifact@v3
