### 2024-01-25 4ec3715
remove dev readme
