+++ CITATION.cff
@@ -68,5 +68,5 @@ keywords:
   - DeepRank
 license: Apache-2.0
 commit: 4e8823758ba03f824b4281f5689cb6a335ab2f6c
-version: "2.1.2"
+version: "3.0.0"
 date-released: "2023-12-21"
