+++ CITATION.cff
@@ -69,4 +69,4 @@ keywords:
 license: Apache-2.0
 commit: 4e8823758ba03f824b4281f5689cb6a335ab2f6c
 version: "3.0.0"
-date-released: "2023-12-21"
+date-released: "2024-01-25"
