### 2024-01-25 8874f86
remove point 6 for making a relese - no more necessary
