### 2024-01-26 c07be6a
lint: select all ruff linting modules in pyproject.toml
individually suppressed those that are unwanted/unrealistic instead of acvtivating each module explicitly
also imporove some annotations in pyproject.toml
