+++ docs/conf.py
@@ -13,14 +13,15 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import configparser  # noqa: F401 (unused-import)
+import configparser
+
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 import os
 import sys
 
-import toml
+import toml  # pyright: ignore[reportMissingModuleSource]
 
 autodoc_mock_imports = [
     "numpy",
