### 2024-01-31 029b0b0
lint: suppress instances of C901 (complexity) and PLR2004 (magic comparisons)
