### 2024-01-31 9eb7484
lint: added missing type annotations throughout
