+++ deeprank2/utils/parsing/patch.py
@@ -20,10 +20,10 @@ class PatchAction:
         self.selection = selection
         self.kwargs = kwargs
 
-    def __contains__(self, key):
+    def __contains__(self, key: str):
         return key in self.kwargs
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: str):
         return self.kwargs[key]
 
 
@@ -33,7 +33,7 @@ class PatchParser:
     ACTION_PATTERN = re.compile(r"^([A-Z]{3,4})\s+([A-Z]+)\s+ATOM\s+([A-Z0-9]{1,3})\s+(.*)$")
 
     @staticmethod
-    def _parse_action_type(s):
+    def _parse_action_type(s: str) -> PatchActionType:
         for type_ in PatchActionType:
             if type_.name == s:
                 return type_
@@ -41,7 +41,7 @@ class PatchParser:
         raise ValueError(f"Unmatched residue action: {s!r}")
 
     @staticmethod
-    def parse(file_):
+    def parse(file_: str) -> list[PatchAction]:
         result = []
         for line in file_:
             if line.startswith(("#", "!")) or len(line.strip()) == 0:
