+++ deeprank2/features/surfacearea.py
@@ -29,10 +29,12 @@ def add_sasa(pdb_path: str, graph: Graph) -> None:
             area = freesasa.selectArea(selection, structure, result)["atom"]
 
         else:
-            raise TypeError(f"Unexpected node type: {type(node.id)}")
+            msg = f"Unexpected node type: {type(node.id)}"
+            raise TypeError(msg)
 
         if np.isnan(area):
-            raise ValueError(f"freesasa returned {area} for {residue}")
+            msg = f"freesasa returned {area} for {residue}"
+            raise ValueError(msg)
         node.features[Nfeat.SASA] = area
 
 
@@ -96,7 +98,8 @@ def add_bsa(graph: Graph) -> None:
             area_key = "atom"
             selection = f"atom, (name {atom.name}) and (resi {atom.residue.number_string}) and (chain {atom.residue.chain.id})"
         else:
-            raise TypeError(f"Unexpected node type: {type(node.id)}")
+            msg = f"Unexpected node type: {type(node.id)}"
+            raise TypeError(msg)
 
     sasa_complete_result = freesasa.calc(sasa_complete_structure)
     sasa_chain_results = {chain_id: freesasa.calc(structure) for chain_id, structure in sasa_chain_structures.items()}
