### 2024-02-03 24618d6
style: remove manual noqa explanations
the full ruff documentation can be accessed by hovering over the suppressed linting rule
