+++ tests/features/test_secondary_structure.py
@@ -36,12 +36,12 @@ def test_secondary_structure_residue() -> None:
         (267, "A", " ", SecondarySctructure.COIL),
         (46, "A", "S", SecondarySctructure.COIL),
         (104, "A", "T", SecondarySctructure.COIL),
-        # (None, '', 'P', SecondarySctructure.COIL),  # not found in test file  # noqa: ERA001 (commented-out code)
+        # (None, '', 'P', SecondarySctructure.COIL),  # not found in test file  # noqa: ERA001
         (194, "A", "B", SecondarySctructure.STRAND),
         (385, "B", "E", SecondarySctructure.STRAND),
         (235, "A", "G", SecondarySctructure.HELIX),
         (263, "A", "H", SecondarySctructure.HELIX),
-        # (0, '', 'I', SecondarySctructure.HELIX),  # not found in test file  # noqa: ERA001 (commented-out code)
+        # (0, '', 'I', SecondarySctructure.HELIX),  # not found in test file  # noqa: ERA001
     ]
 
     for res in residues:
