### 2024-02-03 1ac091b
lint: fix unsafe defaults (B006) and unnecessary Literal type hints (PYI051)
