+++ tests/test_querycollection.py
@@ -18,7 +18,7 @@ from deeprank2.tools.target import compute_ppi_scores
 def _querycollection_tester(
     query_type: str,
     n_queries: int = 3,
-    feature_modules: ModuleType | list[ModuleType] = [components, contact],
+    feature_modules: ModuleType | list[ModuleType] | None = None,
     cpu_count: int = 1,
     combine_output: bool = True,
 ) -> (QueryCollection, str, list[str]):
@@ -35,6 +35,7 @@ def _querycollection_tester(
         combine_output (bool): boolean for combining the hdf5 files generated by the processes.
             By default, the hdf5 files generated are combined into one, and then deleted.
     """
+    feature_modules = feature_modules or [components, contact]
     if query_type == "ppi":
         queries = [
             ProteinProteinInterfaceQuery(
