### 2024-02-03 3cf9dd5
docs: activate missing docstring in public class and public function
- Moved docstrings from __init__ to the class level instead
- Added missing docstrings or suppressed the call for them
- Did not activate missing docstring in public method because there are too many missing
- Deactivated missing docsstrings in tests
