### 2024-02-05 46736db
docs: fix error message structure
