+++ .github/actions/install-python-and-package/action.yml
@@ -57,7 +57,10 @@ runs:
         ### Installing for CPU only on the CI
         conda install pytorch=2.1.1 torchvision=0.16.1 torchaudio=2.1.1 cpuonly=2.0.* -c pytorch
         conda install pyg=2.4.0 -c pyg
-        pip install torch_scatter==2.1.2 torch_sparse==0.6.18 torch_cluster==1.6.3 torch_spline_conv==1.2.2 -f https://data.pyg.org/whl/torch-2.1.0+cpu.html
+        conda install pytorch-scatter=2.1.2 -c pyg
+        conda install pytorch-sparse=0.6.18 -c pyg
+        conda install pytorch-cluster=1.6.3 -c pyg
+        conda install pytorch-spline-conv=1.2.2 -c pyg
     - name: Install dependencies on MacOS
       shell: bash {0}
       env:
