### 2024-02-06 fd5e743
add conda installation of pip deps
