### 2024-02-06 181d6e2
remove pdb2sql from conda installations
