### 2024-02-06 cf19db6
finish adding conda installation for the required deps
