+++ pyproject.toml
@@ -44,15 +44,15 @@ dependencies = [
     "pdb2sql >= 0.5.1",
     # "scikit-learn >= 1.0.2",
     # "chart-studio >= 1.1.0",
-    "biopython >= 1.81",
-    "python-louvain >= 0.16",
-    "markov-clustering >= 0.0.6.dev0",
-    "tqdm >= 4.63.0",
-    "freesasa >= 2.1.0",
-    "tensorboard >= 0.9.0",
-    "protobuf >= 3.20.1",
-    "ruff >= 0.1.13",
-    "dill",
+    # "biopython >= 1.81",
+    # "python-louvain >= 0.16",
+    # "markov-clustering >= 0.0.6.dev0",
+    # "tqdm >= 4.63.0",
+    # "freesasa >= 2.1.0",
+    # "tensorboard >= 0.9.0",
+    # "protobuf >= 3.20.1",
+    # "ruff >= 0.1.13",
+    # "dill",
 ]
 
 [project.optional-dependencies]
