### 2024-02-08 #5983 @gcroci2
Hi @richardjgowers, any news about the publication?
