### 2024-02-08 6d367d7
uniform specifications in conda deps
