### 2024-02-08 ca672f1
insert >= for pytorch and its deps
