### 2024-02-08 b2e4c27
try to do conda installations with yml file
