+++ env/requirements2.txt
@@ -1,0 +1,1 @@
+pdb2sql>=0.5.1
