+++ .github/actions/install-python-and-package/action.yml
@@ -49,16 +49,17 @@ runs:
       if: runner.os == 'Linux'
       run: |
         # Install deeprank2 conda dependencies
-        ## sbl channel
-        conda install dssp>=4.2.2.1 -c sbl
-        ## bioconda channel
-        conda install msms>=2.6.1 markov_clustering>=0.0.6 -c bioconda
-        ## pytorch and pyg channels
-        ### Installing for CPU only on the CI
-        conda install pytorch=2.1.1 torchvision>=0.16.1 torchaudio>=2.1.1 cpuonly>=2.0 -c pytorch
-        conda install pyg>=2.4.0 pytorch-scatter>=2.1.2 pytorch-sparse>=0.6.18 pytorch-cluster>=1.6.3 pytorch-spline-conv>=1.2.2 -c pyg
-        ## conda-forge channel
-        conda install tables>=3.8.0 numpy>=1.21.5 scipy>=1.11.2 h5py>=3.6.0 networkx>=2.6.3 matplotlib>=3.5.1 scikit-learn>=1.0.2 chart-studio>=1.1.0 biopython>=1.81 python-louvain>=0.16 tqdm>=4.63.0 freesasa>=2.1.0 tensorboard>=0.9.0 protobuf>=3.20.1 ruff>=0.1.13 dill>=0.3.8 -c conda-forge
+        conda env update -f env/env2.yml
+        # ## sbl channel
+        # conda install dssp>=4.2.2.1 -c sbl
+        # ## bioconda channel
+        # conda install msms>=2.6.1 markov_clustering>=0.0.6 -c bioconda
+        # ## pytorch and pyg channels
+        # ### Installing for CPU only on the CI
+        # conda install pytorch=2.1.1 torchvision>=0.16.1 torchaudio>=2.1.1 cpuonly>=2.0 -c pytorch
+        # conda install pyg>=2.4.0 pytorch-scatter>=2.1.2 pytorch-sparse>=0.6.18 pytorch-cluster>=1.6.3 pytorch-spline-conv>=1.2.2 -c pyg
+        # ## conda-forge channel
+        # conda install tables>=3.8.0 numpy>=1.21.5 scipy>=1.11.2 h5py>=3.6.0 networkx>=2.6.3 matplotlib>=3.5.1 scikit-learn>=1.0.2 chart-studio>=1.1.0 biopython>=1.81 python-louvain>=0.16 tqdm>=4.63.0 freesasa>=2.1.0 tensorboard>=0.9.0 protobuf>=3.20.1 ruff>=0.1.13 dill>=0.3.8 -c conda-forge
     - name: Install dependencies on MacOS
       shell: bash {0}
       env:
