+++ pyproject.toml
@@ -41,7 +41,7 @@ dependencies = [
     # "h5py >= 3.6.0",
     # "networkx >= 2.6.3",
     # "matplotlib >= 3.5.1",
-    "pdb2sql >= 0.5.1",
+    # "pdb2sql >= 0.5.1",
     # "scikit-learn >= 1.0.2",
     # "chart-studio >= 1.1.0",
     # "biopython >= 1.81",
