+++ .github/actions/install-python-and-package/action.yml
@@ -26,30 +26,36 @@ runs:
       uses: styfle/cancel-workflow-action@0.4.0
       with:
         access_token: ${{ github.token }}
+
     - uses: actions/checkout@v3
-    - name: Setup conda
-      uses: s-weigand/setup-conda@v1
+
+    - name: Setup miniconda
+      uses: conda-incubator/setup-miniconda@v2
       with:
-        update-conda: true
+        auto-update-conda: true
+        miniforge-variant: Mambaforge
+        channels: conda-forge
         python-version: ${{ inputs.python-version }}
-        conda-channels: pytorch, pyg, bioconda, defaults, sbl, conda-forge
+        activate-environment: deeprank2
+        environment-file: env/env2.yml
+        use-mamba: true
+
     - run: |
         conda --version
         conda env list
-      shell: bash {0}
+
     - name: Python info
-      shell: bash -e {0}
       run: |
         which python3
         python3 --version
-    - name: Install dependencies on Linux
-      shell: bash {0}
+
+    - name: Activate deeprank2 environment on Linux
       env:
         CMAKE_INSTALL_PREFIX: .local
       if: runner.os == 'Linux'
       run: |
-        # Install deeprank2 conda dependencies
-        conda env update -f env/env2.yml
+        conda activate deeprank2
+        # # Install deeprank2 conda dependencies
         # ## sbl channel
         # conda install dssp>=4.2.2.1 -c sbl
         # ## bioconda channel
@@ -60,36 +66,12 @@ runs:
         # conda install pyg>=2.4.0 pytorch-scatter>=2.1.2 pytorch-sparse>=0.6.18 pytorch-cluster>=1.6.3 pytorch-spline-conv>=1.2.2 -c pyg
         # ## conda-forge channel
         # conda install tables>=3.8.0 numpy>=1.21.5 scipy>=1.11.2 h5py>=3.6.0 networkx>=2.6.3 matplotlib>=3.5.1 scikit-learn>=1.0.2 chart-studio>=1.1.0 biopython>=1.81 python-louvain>=0.16 tqdm>=4.63.0 freesasa>=2.1.0 tensorboard>=0.9.0 protobuf>=3.20.1 ruff>=0.1.13 dill>=0.3.8 -c conda-forge
-    - name: Install dependencies on MacOS
-      shell: bash {0}
-      env:
-        CMAKE_INSTALL_PREFIX: .local
-      if: runner.os == 'macOS'
-      run: |
-        # Install dependencies not handled by setuptools
-        ## DSSP
-        conda install -c sbl dssp>=4.2.2.1
-        ## MSMS
-        cd /tmp/
-        wget http://mgltools.scripps.edu/downloads/tars/releases/MSMSRELEASE/REL2.6.1/msms_i86Linux2_2.6.1.tar.gz
-        sudo mkdir /usr/local/lib/msms
-        cd /usr/local/lib/msms
-        sudo tar zxvf /tmp/msms_i86Linux2_2.6.1.tar.gz
-        sudo ln -s /usr/local/lib/msms/msms.i86Linux2.2.6.1 /usr/local/bin/msms
-        sudo ln -s /usr/local/lib/msms/pdb_to_xyzr* /usr/local/bin
-        ## PyTorch, PyG, PyG adds
-        ### Installing for CPU only on the CI
-        conda install pytorch torchvision torchaudio cpuonly -c pytorch
-        pip install torch_geometric
-        pip install torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-$(python3 -c "import torch; print(torch.__version__)")+cpu.html
-        # PyTables via conda only for MacOS
-        conda install pytables
+
     - name: Install the GitHub repository version of the package
[…]
