+++ .github/workflows/build-repo.yml
@@ -39,14 +39,22 @@ jobs:
         os: ["ubuntu-latest"]
         python-version: ["3.10"] # ["3.10", "3.11"]
 
+    # https://github.com/marketplace/actions/setup-miniconda#use-a-default-shell
+    defaults:
+      run:
+        shell: bash -l {0}
+
     steps:
       - uses: actions/checkout@v3
+
       - uses: ./.github/actions/install-python-and-package
         with:
           python-version: ${{ matrix.python-version }}
           extras-require: test, publishing
           pkg-installation-type: "repository"
+
       - name: Run unit tests
         run: pytest -v
+
       - name: Verify that we can build the package
         run: python3 -m build
