### 2024-02-08 489d9bc
add bash property
