### 2024-02-08 edcf798
try to fix env activation
