### 2024-02-08 03461dc
try to fix again conda env activation
