+++ .github/actions/install-python-and-package/action.yml
@@ -71,7 +71,7 @@ runs:
         # conda install tables>=3.8.0 numpy>=1.21.5 scipy>=1.11.2 h5py>=3.6.0 networkx>=2.6.3 matplotlib>=3.5.1 scikit-learn>=1.0.2 chart-studio>=1.1.0 biopython>=1.81 python-louvain>=0.16 tqdm>=4.63.0 freesasa>=2.1.0 tensorboard>=0.9.0 protobuf>=3.20.1 ruff>=0.1.13 dill>=0.3.8 -c conda-forge
 
     - name: Install the GitHub repository version of the package
-      shell: bash {0}
+      shell: bash -el {0}
       if: ${{ inputs.pkg-installation-type == 'repository' }}
       run: |
         conda activate deeprank2
