+++ env/env2.yml
@@ -7,7 +7,8 @@ channels:
   - conda-forge
   - sbl
 dependencies:
-  - dssp>=4.2.2.1
+  - sbl::libcifpp>=5.1.0
+  - sbl::dssp>=4.2.2.1
   - msms>=2.6.1
   - markov_clustering>=0.0.6
   - pytorch=2.1.1
