### 2024-02-13 8f66c6d
remove deps from pyproject
