+++ .github/workflows/coveralls.yml
@@ -39,6 +39,11 @@ jobs:
         os: ["ubuntu-latest"]
         python-version: ["3.10"]
 
+    # https://github.com/marketplace/actions/setup-miniconda#use-a-default-shell
+    defaults:
+      run:
+        shell: bash -l {0}
+
     steps:
       - uses: actions/checkout@v3
       - uses: ./.github/actions/install-python-and-package
