+++ env/deeprank2-py310.yml
@@ -7,6 +7,9 @@ channels:
   - conda-forge
   - sbl
 dependencies:
+  - python==3.10
+  - pip>=23.3
+  - notebook>=7.0.6
   - sbl::libcifpp>=5.1.0
   - sbl::dssp>=4.2.2.1
   - msms>=2.6.1
@@ -36,5 +39,6 @@ dependencies:
   - protobuf>=3.20.1
   - ruff>=0.1.13
   - dill>=0.3.8
+  - pyarrow>=15.0.0
   - pip:
-      - --requirement requirements2.txt
+      - --requirement requirements.txt
