+++ env/requirements-docker.txt
@@ -1,0 +1,2 @@
+pdb2sql>=0.5.1
+deeprank2>=3.0.0
