+++ env/deeprank2-docker.yml
@@ -41,4 +41,4 @@ dependencies:
   - dill>=0.3.8
   - pyarrow>=15.0.0
   - pip:
-      - --requirement requirements.txt
+      - --requirement requirements-docker.txt
