### 2024-02-15 19f8313
update docker file with new env files names
