### 2024-02-15 94b4194
update action.yml with the renamed env file
