### 2024-02-19 6ab611d
add pdb2sql conda installation and remove the pip installation
