+++ env/requirements.txt
@@ -1,1 +1,0 @@
-pdb2sql>=0.5.1
