+++ env/deeprank2-docker.yml
@@ -32,6 +32,7 @@ dependencies:
   - scikit-learn>=1.0.2
   - chart-studio>=1.1.0
   - biopython>=1.81
+  - pdb2sql>=0.5.1
   - python-louvain>=0.16
   - tqdm>=4.63.0
   - freesasa>=2.1.0
