### 2024-02-19 f10336a
update docs for new installation
