+++ docs/installation.md
@@ -4,8 +4,8 @@
 - [Installation](#installation)
   - [Containerized Installation](#containerized-installation)
   - [Local/remote installation](#localremote-installation)
-      - [YML file installation](#yml-file-installation)
-    - [Manual installation](#manual-installation)
+    - [YML file installation (recommended)](#yml-file-installation-recommended)
+    - [Manual installation (not recommended)](#manual-installation-not-recommended)
     - [Testing DeepRank2 installation](#testing-deeprank2-installation)
 - [Contributing](#contributing)
 
@@ -46,7 +46,7 @@ Local installation is formally only supported on the latest stable release of ub
 
 Before installing DeepRank2 please ensure you have [GCC](https://gcc.gnu.org/install/) installed: if running `gcc --version` gives an error, run `sudo apt-get install gcc`.
 
-#### YML file installation
+#### YML file installation (recommended)
 
 You can use the provided YML file for creating a [conda environment](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html) containing the latest stable release of DeepRank2 and all its dependencies.
 This will install the CPU-only version of DeepRank2 on Python 3.10.
@@ -60,27 +60,28 @@ cd deeprank2
 # Ensure you are in your base environment
 conda activate
 # Create the environment
-conda env create -f env/environment.yml
+conda env create -f env/deeprank2.yml
 # Activate the environment
 conda activate deeprank2
+# Install the latest deeprank2 release
+pip install deeprank2
 ```
 
 See instructions below to [test](#testing-deeprank2-installation) that the installation was succesful.
 
-### Manual installation
+#### Manual installation (not recommended)
 
-If you want to use the GPUs, choose a specific python version, are a MacOS user, or if the YML installation was not succesful, you can install the package manually. We advise to do this inside a [conda virtual environment](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html).
-If you have any issues during installation of dependencies, please refer to the official documentation for each package (linked below), as our instructions may be out of date (last tested on 19 Jan 2024):
+If you want to use the GPUs, choose a specific python version, are a MacOS user, or if the YML installation was not successful, you can install the package manually. We advise to do this inside a [conda virtual environment](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html).
 
-- [DSSP 4](https://anaconda.org/sbl/dssp): `conda install -c sbl dssp`
-- [MSMS](https://anaconda.org/bioconda/msms): `conda install -c bioconda msms`
-  - [Here](https://ssbio.readthedocs.io/en/latest/instructions/msms.html) for MacOS with M1 chip users.
-- [PyTorch](https://pytorch.org/get-started/locally/): `conda install pytorch torchvision torchaudio cpuonly -c pytorch`
-  - Pytorch regularly publishes updates and not all newest versions will work stably with DeepRank2. Currently, the package is tested using [PyTorch 2.1.1](https://pytorch.org/get-started/previous-versions/#v211).
+You can first remove from `env/deeprank2.yml` the packages that cannot be installed properly, or the ones that you want to install differently (e.g., pytorch-related packages if you wish to install the CUDA version), and then proceed with the environment creation by using the edited YML file: `conda env create -f env/deeprank2.yml`. Then activate the environment, and proceed with installing the missing packages, which might fall into the following list. If you have any issues during installation of dependencies, please refer to the official documentation for each package (linked below), as our instructions may be out of date (last tested on 19 Feb 2024):
+
+- [MSMS](https://anaconda.org/bioconda/msms): [Here](https://ssbio.readthedocs.io/en/latest/instructions/msms.html) for MacOS with M1 chip users.
[…]
