+++ paper/paper.md
@@ -47,9 +47,9 @@ authors:
     orcid: 0000-0002-2613-538X
     affiliation: 2
 affiliations:
-  - name: Netherlands eScience Center, Amsterdam, NL
+  - name: Netherlands eScience Center, Amsterdam, The Netherlands
     index: 1
-  - name: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, NL
+  - name: Center for Molecular Bioinformatics, Radboud Universitair Medisch Centrum, Nijmegen, Gelderland, The Netherlands
     index: 2
   - name: Independent Researcher
     index: 3
