### 2024-02-20 #5983 @gcroci2
> @gcroci2 as AEiC I will now help to process this work for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order. I only have the below points that require your attention:
>
> * [ ]  Please spell out `NL` as `The Netherlands` in your affiliations.
> * [ ]  Add this DOI: [10.48550/arXiv.2202.08320](https://doi.org/10.48550/arXiv.2202.08320) for the `TorchDrug` reference.
> * [ ]  Add this DOI: [10.1002/(SICI)1097-0134(1997)1+%3C50::AID-PROT8%3E3.0.CO;2-S](https://doi.org/10.1002/(SICI)1097-0134(1997)1+%3C50::AID-PROT8%3E3.0.CO;2-S) for `Evaluation of comparative protein structure modeling by MODELLER-3`, also, ensure you remove the text `Google Scholar There Is No Corresponding Record for This Reference.` from your .bib file for this one.
> * [ ]  Add this DOI: [10.1038/s41586-023-05993-x](https://doi.org/10.1038/s41586-023-05993-x) for `De novo design of protein interactions with learned surface fingerprints`
> * [ ]  Add this DOI: [10.1126/science.abj8754](https://doi.org/10.1126/science.abj8754) for `Accurate prediction of protein structures and interactions using a three-track neural network.`.
>
> There is no need to update the archive/version for these minor paper changes.

All done! @Kevin-Mattheus-Moerman
