+++ tests/test_trainer.py
@@ -1,5 +1,4 @@
 import glob
-import inspect
 import logging
 import os
 import shutil
@@ -8,7 +7,6 @@ import unittest
 import warnings
 
 import h5py
-import numpy as np
 import pandas as pd
 import pytest
 import torch
