+++ CITATION.cff
@@ -68,5 +68,5 @@ keywords:
   - DeepRank
 license: Apache-2.0
 commit: 4e8823758ba03f824b4281f5689cb6a335ab2f6c
-version: "3.0.0"
+version: "3.0.1"
 date-released: "2024-01-25"
