### 2024-02-22 27edea2
update deeprank2 minimum version for docker
