+++ README.md
@@ -45,7 +45,7 @@ Main features:
     - [Containerized Installation](#containerized-installation)
     - [Local/remote installation](#localremote-installation)
       - [YML file installation (recommended)](#yml-file-installation-recommended)
-      - [Manual installation (not recommended)](#manual-installation-not-recommended)
+      - [Manual installation (customizable)](#manual-installation-customizable)
       - [Testing DeepRank2 installation](#testing-deeprank2-installation)
   - [Contributing](#contributing)
   - [Using DeepRank2](#using-deeprank2)
