### 2024-02-27 #5983 @Kevin-Mattheus-Moerman
@openjournals/dev the above DOI is actually correct. I think our bot doesn't handle the special characters in this odd DOI very well.
